Calculate the alpha diversity of each sample.

Calculate the alpha diversity of each sample.

Usage

adiv_table(biom, rarefy = FALSE, adiv = "Shannon", long = FALSE, md = FALSE)

Arguments

biom

A matrix, simple_triplet_matrix, or BIOM object, as returned from read_biom(). For matrices, the rows and columns are assumed to be the taxa and samples, respectively.

rarefy

Control how/whether rarefactions are done prior to alpha diversity computations. Options are:

FALSE - Use each sample's current set of observations without applying any rarefaction.

TRUE - Automatically select and apply a single rarefaction.

"multi", "multi_log", "multi_even" - Automatically select and apply multiple rarefactions. "multi" provides "multi_log" at the low end and "multi_even" at the high end.

integer vector - Rarefy at the specified depth(s).

Default: FALSE

adiv

One or more alpha diversity measures to use. Options are: "OTUs", "Shannon", "Chao1", "Simpson", and/or "InvSimpson". Non-ambiguous abbreviations are also accepted. You can use adiv="all" as shortcut for adiv=c("OTUs", "Shannon", "Chao1", "Simpson", "InvSimpson") Default: "Shannon".

long

Pivot the returned data to long format?

FALSE - Each alpha diversity metric always has its own column, named ".OTUs", ".Shannon", etc.

TRUE - The name of the metric is in an ".adiv" column, and the alpha diversity values are in a ".diversity" column.

Default: FALSE

md

Include metadata in the output data frame? Options are:

FALSE - Don't include metadata.

TRUE - Include all metadata.

character vector - Include only the specified metadata columns.

Default: FALSE

Value

A data frame of diversity values for each sample in biom. The first column name is .sample. Remaining column names are dependent on the function's arguments.

Examples

    library(rbiom)
    
    biom <- sample_select(hmp50, 1:6)
    
    adiv_table(biom, adiv="all")
#>   .sample .OTUs .Shannon    .Chao1  .Simpson .InvSimpson
#> 1   HMP01    49 1.741153  59.08333 0.5668414    2.308623
#> 2   HMP02    75 2.587403 104.38889 0.8133914    5.358811
#> 3   HMP03    75 2.950982  90.04167 0.8936622    9.403996
#> 4   HMP04    83 3.255785  93.32143 0.9323645   14.785121
#> 5   HMP05    67 1.462651 127.75000 0.5252255    2.106263
#> 6   HMP06   105 3.087615 131.88889 0.9103916   11.159668
    adiv_table(biom, long = TRUE, md = TRUE)
#>   .sample .depth   .adiv .diversity Age BMI     Body Site    Sex
#> 1   HMP01   1660 Shannon   1.741153  22  20 Buccal mucosa Female
#> 2   HMP02   1371 Shannon   2.587403  24  23 Buccal mucosa   Male
#> 3   HMP03   1353 Shannon   2.950982  28  26        Saliva   Male
#> 4   HMP04   1895 Shannon   3.255785  25  23        Saliva   Male
#> 5   HMP05   3939 Shannon   1.462651  27  24 Buccal mucosa Female
#> 6   HMP06   4150 Shannon   3.087615  32  25        Saliva   Male