Estimate the diversity of each sample.

Estimate the diversity of each sample.

alpha.div(biom, progressbar = FALSE)

Arguments

biom
A matrix, simple_triplet_matrix, or BIOM object, as returned from read.biom. For matrices, the rows and columns are assumed to be the taxa and samples, respectively.

progressbar
Whether to display a progress bar and status messages (TRUE/FALSE). Will automatically tie in with shiny if run within a shiny session.

Value

A numeric matrix of four diversity values for each sample in biom. Row names are the sample names and column names are the diversity metrics: OTUs, Shannon, Simpson, and Chao1.

Examples

    library(rbiom)

    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)

    ad <- alpha.div(biom)
    head(ad)

#>       OTUs  Shannon   Simpson     Chao1
#> HMP01   89 2.972967 0.9085382 123.72222
#> HMP02   41 2.223793 0.8243218  43.08333
#> HMP03   35 1.601946 0.6244216  63.16667
#> HMP04   49 1.310901 0.4622226  70.33333
#> HMP05   54 2.189982 0.7706529  79.78571
#> HMP06   44 1.250978 0.5103010  62.06250