Display distances between samples as a 2-D scatter plot.

Display distances between samples as a 2-D scatter plot.

Usage

bdiv_biplot(
  biom,
  metric = "Bray-Curtis",
  ord = "PCoA",
  layers = NULL,
  color.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  limit.by = NULL,
  weighted = TRUE,
  rank = NULL,
  taxa = 5,
  p.top = Inf,
  p.adj = "fdr",
  perms = 1000,
  ...
)

Arguments

biom

A BIOM object, as returned from read_biom.

metric

Beta diversity metric to use for calculating inter-sample distances. Options are: "Bray-Curtis", "Manhattan", "Euclidean", "Jaccard", or "UniFrac". Default: "Bray-Curtis".

ord

Ordination method to use for condensing n-dimensional data to two dimensions. Options are: "UMAP", "tSNE", "NMDS", or "PCoA". Default: "PCoA".

layers

What graphical elements to use for drawing. Options are: point, spider, ellipse, and name for samples, and mean, taxon, and arrow for biplots. Single letter abbreviations are also accepted. For instance, c("point", "ellipse") is equivalent to c("p", "e") and "pe". See vignette("biplots") for examples of each. Default: "pce"/"p" for ordinations with/without a color.by argument.

color.by, shape.by, facet.by, limit.by

Metadata columns to use for data partitioning. Default: NULL

weighted

When employing a beta diversity metric, use the weighted version. Default: TRUE.

rank

What rank of taxa to display, for example "Phylum" or "Genus". Run taxa_ranks() to see all options for a given BIOM object. The default, NULL, selects the lowest level.

taxa

Which taxa to display. An integer value will show the top n most abundant taxa. A value 0 <= n < 1 will show any taxa with that mean abundance or greater (e.g. 0.1). A character vector of taxon names will show only those taxa. Default: 5.

p.top

Only display taxa with the most significant differences in abundance. If p.top is >= 1, then the p.top most significant taxa are displayed. If p.top is less than one, all taxa with an adjusted p-value <= p.top are displayed. Recommended to be used in combination with the taxa parameter to set a lower bound on the mean abundance of considered taxa. Default: Inf.

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: "fdr".

perms

Number of random permutations to use for estimating statistical significance. Default: 1000.

...

Parameters for underlying functions. Prefixing parameter names with a layer name ensures that a particular parameter is passed to, and only to, that layer. For instance, dot.size = 2 or d.size = 2 ensures only the dotplot layer has its size set to 2.

Value

A ggplot2 plot. The computed data points and statistics will be attached as attr(p, 'data') and attr(p, 'stats'), respectively.

See also

Examples

    library(rbiom)
    
    biom <- rarefy(hmp50) 
    bdiv_biplot(biom, color.by="Body Site", rank="Genus")
#> Error in complete.cases(new_md): no input has determined the number of cases