Define sample PAM clusters from beta diversity.

Define sample PAM clusters from beta diversity.

Usage

bdiv_clusters(
  biom,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  k = 5,
  ...
)

Arguments

biom

An rbiom object, such as from as_rbiom(). Any value accepted by as_rbiom() can also be given here.

bdiv

Beta diversity distance algorithm(s) to use. Options are: "Bray-Curtis", "Manhattan", "Euclidean", "Jaccard", and "UniFrac". For "UniFrac", a phylogenetic tree must be present in biom or explicitly provided via tree=. Default: "Bray-Curtis"

Multiple/abbreviated values allowed.

weighted

Take relative abundances into account. When weighted=FALSE, only presence/absence is considered. Default: TRUE

Multiple values allowed.

tree

A phylo object representing the phylogenetic relationships of the taxa in biom. Only required when computing UniFrac distances. Default: biom$tree

k

Number of clusters. Default: 5L

...

Passed on to cluster::pam().

Value

A numeric factor assigning samples to clusters.

See also

Examples

    library(rbiom)
    
    biom <- rarefy(hmp50)
    biom$metadata$bray_cluster <- bdiv_clusters(biom)
#> Error in bdiv_clusters(biom): could not find function "bdiv_clusters"
    
    pull(biom, 'bray_cluster')[1:10]
#> Error in validate_var_choices("xi", choices, evar = evar): Invalid `var` option(s): bray_cluster
#> Choices are: .sample, Age, BMI, Body Site, Sex
    
    bdiv_ord_plot(biom, stat.by = "bray_cluster")
#> Error in validate_var_choices("xi", choices, evar = evar): Invalid `stat.by` option(s): bray_cluster
#> Choices are: .sample, Age, BMI, Body Site, Sex