Calculate PCoA and other ordinations, including taxa biplots and statistics.

Calculate PCoA and other ordinations, including taxa biplots and statistics.

Usage

bdiv_ord_table(
  biom,
  bdiv = "Bray-Curtis",
  ord = "UMAP",
  weighted = TRUE,
  md = NULL,
  k = 2,
  split.by = NULL,
  stat.by = NULL,
  tree = NULL,
  test = "adonis2",
  seed = 0,
  permutations = 999,
  rank = NULL,
  taxa = 5,
  p.top = Inf,
  p.adj = "fdr",
  unc = "singly",
  ...
)

Arguments

biom

A BIOM object, as returned from read_biom(). Alternatively, a dist class distance matrix can be given, in which case only the parameters ord, k, and ... are allowed.

bdiv

The beta diversity distance algorithm to use. Options are: "Bray-Curtis", "Manhattan", "Euclidean", "Jaccard", and "UniFrac". A phylogenetic tree must be present in biom or explicitly provided via tree= to use the UniFrac methods. Multiple values allowed. Default: "Bray-Curtis".

ord

Method for reducing dimensionality. Options are:

"UMAP" - Uniform manifold approximation and projection; uwot::umap().

"PCoA" - Principal coordinate analysis; ape::pcoa().

"NMDS" - Nonmetric multidimensional scaling; vegan::metaMDS().

"tSNE" - t-distributed stochastic neighbor embedding; tsne::tsne().

Default: "UMAP"

Multiple values allowed. Non-ambiguous abbreviations are allowed.

weighted

Take relative abundances into account. When weighted=FALSE, only presence/absence is considered. Multiple values allowed. Default: TRUE

md

Include metadata in the output data frame? Options are:

NULL - Don't include metadata. (Default)

TRUE - Include all metadata.

character vector - Include only the specified metadata columns.

Default: NULL

k

Number of ordination dimensions to return. Either 2L or 3L. Default: 2L

split.by

Name(s) of metadata columns that the data should be split by prior to calculating distance matrices, ordinations, and statistics. Equivalent to the facet.by parameter in bdiv_ord_plot(). Default: NULL

stat.by

Name of the metadata column for statistical groups. Equivalent to the color.by parameter in bdiv_ord_plot(). Default: NULL

tree

A phylo object representing the phylogenetic relationships of the taxa in biom. Will be taken from the tree defined in biom if not explicitly specified. Only required when computing UniFrac distance matrices. Default: NULL

test, seed, permutations

Passthrough parameters to distmat_stats() for assessing significance of stat.by and rank groups. Default: test="adonis2", seed=0, permutations=999

rank

Biplot What rank of taxa to use for the biplot (e.g. "Phylum"), or NULL for no biplot. Run taxa_ranks() to see all options for a given BIOM object. Default: NULL

taxa

Biplot Which taxa to include. An integer value will return the top n most abundant taxa. A value 0 <= n < 1 will return any taxa with that mean abundance or greater (e.g. 0.1). A character vector of taxon names will show only those taxa. Default: 5

p.top

Biplot Only return taxa with the most significant differences in abundance. If p.top is >= 1, then the p.top most significant taxa are displayed. If p.top is less than one, all taxa with an adjusted p-value <= p.top are displayed. Recommended to be used in combination with the taxa parameter to set a lower bound on the mean abundance of considered taxa. Default: Inf

p.adj

Biplot Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: fdr

unc

Biplot How to handle unclassified, uncultured, and similarly ambiguous taxa names. Options are:

"singly" - Replaces them with the OTU name.

"grouped" - Replaces them with a higher rank's name.

"drop" - Excludes them from the result.

"asis" - To not check/modify any taxa names.

Default: "singly"

Non-ambiguous abbreviations are allowed.

...

Additional arguments to pass on to uwot::umap(), ape::pcoa(), vegan::metaMDS(), or tsne::tsne().

Value

A data.frame with columns .sample, .ord, .x, .y, and (optionally) .z.

    If \code{biom} is a \code{BIOM} object, then \code{.weighted}, 
    \code{.bdiv}, and any columns given by \code{md}, \code{split.by}, 
    and \code{stat.by} are included as well.
    
    If \code{stat.by} is given, then \code{attr(, 'sample_stats')}
    and \code{attr(, 'sample_stats_cmds')} are set.
    
    If \code{rank} is given, then \code{attr(, 'taxa_coords')},
    \code{attr(, 'taxa_stats')}, and \code{attr(, 'taxa_stats_cmds')} 
    are set.

See also

Examples

    library(rbiom) 
    
    ord <- bdiv_ord_table(hmp50, "bray", "pcoa", stat.by="Body Site", rank="g")
    head(ord)
#>   .weighted       .bdiv .ord .sample     .x       .y     Body Site
#> 1      TRUE Bray-Curtis PCoA   HMP01 -0.387 -0.00895 Buccal mucosa
#> 2      TRUE Bray-Curtis PCoA   HMP02 -0.461 -0.01350 Buccal mucosa
#> 3      TRUE Bray-Curtis PCoA   HMP03 -0.407 -0.01430        Saliva
#> 4      TRUE Bray-Curtis PCoA   HMP04 -0.378 -0.01230        Saliva
#> 5      TRUE Bray-Curtis PCoA   HMP05 -0.452 -0.01140 Buccal mucosa
#> 6      TRUE Bray-Curtis PCoA   HMP06 -0.411 -0.01230        Saliva
    attr(ord, 'sample_stats')
#>   .weighted       .bdiv .stat   .z .p.val .adj.p        .test
#> 1      TRUE Bray-Curtis  19.2 72.2  0.001  0.001 adonis2; fdr
    attr(ord, 'taxa_stats')
#>   .weighted       .bdiv .rank             .taxa .stat   .z .p.val .adj.p
#> 1      TRUE Bray-Curtis Genus     Lactobacillus  8.62 16.8  0.001  0.001
#> 2      TRUE Bray-Curtis Genus     Streptococcus 10.10 19.5  0.001  0.001
#> 3      TRUE Bray-Curtis Genus       Bacteroides  8.72 16.7  0.001  0.001
#> 4      TRUE Bray-Curtis Genus Corynebacterium 1  8.53 16.0  0.001  0.001
#> 5      TRUE Bray-Curtis Genus       Haemophilus  6.56 12.3  0.001  0.001
#>          .test
#> 1 adonis2; fdr
#> 2 adonis2; fdr
#> 3 adonis2; fdr
#> 4 adonis2; fdr
#> 5 adonis2; fdr