Test beta diversity vs categorical or numeric metadata.

Test beta diversity vs categorical or numeric metadata.

Usage

bdiv_stats(
  biom,
  stat.by,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  test = "adonis2",
  within = NULL,
  between = NULL,
  split.by = NULL,
  seed = 0,
  permutations = 999,
  p.adj = "fdr"
)

Arguments

biom

An rbiom-class object, or data coercible with as_rbiom().

stat.by

The categorical or numeric metadata field over which statistics should be calculated. Required.

bdiv

Beta diversity distance algorithm(s) to use. Options are: "Bray-Curtis", "Manhattan", "Euclidean", "Jaccard", and "UniFrac". For "UniFrac", a phylogenetic tree must be present in biom or explicitly provided via tree=. Default: "Bray-Curtis"

Multiple/abbreviated values allowed.

weighted

Take relative abundances into account. When weighted=FALSE, only presence/absence is considered. Default: TRUE

Multiple values allowed.

tree

A phylo object representing the phylogenetic relationships of the taxa in biom. Only required when computing UniFrac distances. Default: otu_tree(biom)

test

Permutational test for accessing significance. Options are:

"adonis2" - Permutational MANOVA; vegan::adonis2().

"mrpp" - Multiple response permutation procedure; vegan::mrpp().

Default: "adonis2"

Abbreviations are allowed.

within, between

Metadata field(s) for intra- or inter- sample comparisons. Default: NULL

split.by

Name(s) of metadata columns that the data should be split by prior to any calculations. Default: NULL

seed

Random seed for permutations. Default: 0

permutations

Number of random permutations to use. Default: 999

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: "fdr".

Value

A tibble data frame with summary statistics.

The R code or generating these statistics is in $code, and the object history is in $history.

See also

Examples

    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    
    bdiv_stats(biom, stat.by = "Body Site", split.by = "Sex")
#> # Test:     vegan::adonis2(~ `Body Site`, permutations = 999)
#> # A tibble: 2 × 8
#>   Sex    .weighted .bdiv          .n .stat    .z .p.val .adj.p
#>   <fct>  <lgl>     <chr>       <int> <dbl> <dbl>  <dbl>  <dbl>
#> 1 Female TRUE      Bray-Curtis    30  10.6  35.3  0.001  0.001
#> 2 Male   TRUE      Bray-Curtis    19  12.1  30.5  0.001  0.001
      
    bdiv_stats(biom, stat.by = "BMI", bdiv = c("bray", "unifrac"))
#> # Test:     vegan::adonis2(~ BMI, permutations = 999)
#> # A tibble: 2 × 7
#>   .weighted .bdiv          .n .stat      .z .p.val .adj.p
#>   <lgl>     <chr>       <int> <dbl>   <dbl>  <dbl>  <dbl>
#> 1 TRUE      UniFrac        49 1.80   1.22    0.106  0.212
#> 2 TRUE      Bray-Curtis    49 0.973 -0.0278  0.431  0.431
    
    # The R code used to compute the stats is in $code.
    tbl <- bdiv_stats(biom, stat.by = "Sex")
    tbl$code
#> dm <- bdiv_distmat(biom = biom, bdiv = 'Bray-Curtis', weighted = TRUE)
#> grouping <- sample_metadata(biom, "Sex")[attr(dm, 'Labels')]
#> 
#> set.seed(0)
#> 
#> vegan::adonis2(formula = dm ~ grouping, permutations = 999) %>%
#>   vegan::permustats() %>%
#>   summary() %>%
#>   with(data.frame(.stat = statistic, .z = z, .p.val = p)) %>%
#>   tryCatch(
#>     error   = function (e) data.frame(.stat=NA, .z=NA, .p.val=NA), 
#>     warning = function (w) data.frame(.stat=NA, .z=NA, .p.val=NA) ) %>%
#>   data.frame(row.names = NULL, .n = attr(dm, 'Size'), .) %>%
#>   data.frame(.weighted = TRUE, .bdiv = "Bray-Curtis", .) %>%
#>   transform(.adj.p = stats::p.adjust(.p.val, 'fdr')) %>%
#>   plyr::arrange(.p.val)