Make a data.frame of distances between samples.

Make a data.frame of distances between samples.

Usage

bdiv_table(
  biom,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  md = FALSE,
  within = NULL,
  between = NULL
)

Arguments

biom

A matrix, simple_triplet_matrix, or BIOM object (e.g. as returned from read_biom()). For matrices, the rows and columns are assumed to be the taxa and samples, respectively.

bdiv

The beta diversity distance algorithm to use. Options are: "Bray-Curtis", "Manhattan", "Euclidean", "Jaccard", and "UniFrac". Non-ambiguous abbreviations of the algorithm names are also accepted. For "UniFrac", a phylogenetic tree must be present in biom or explicitly provided via tree=. Default: "Bray-Curtis".

weighted

Take relative abundances into account. When weighted=FALSE, only presence/absence is considered. Default: TRUE

tree

A phylo object representing the phylogenetic relationships of the taxa in biom. Will be taken from the tree defined in biom if not explicitly specified. Only required when computing UniFrac distance matrices. Default: NULL

md

Include metadata in the output data frame? Options are:

FALSE - Don't include metadata.

TRUE - Include all metadata.

character vector - Include only the specified metadata columns. Column names can be prefixed with == or != to indicate that only within or between groupings, respectively, are to be kept. See examples below.

Default: FALSE

within, between

Metadata column name(s) for intra- or inter- sample comparisons. Default: within=NULL, between=NULL

Value

A data.frame with first three columns named ".sample1", ".sample2", and ".distance".

See also

Examples

    library(rbiom)
    
    # Return in long format with metadata
    biom <- sample_select(hmp50, 18:21)
    bdiv_table(biom, 'unifrac', md = c("Body Site", "Sex"))
#>    .sample1 .sample2 .distance       Body Site            Sex
#> 5     HMP18    HMP19 0.6651627 Saliva vs Stool Female vs Male
#> 9     HMP18    HMP20 0.6810017 Saliva vs Stool Female vs Male
#> 10    HMP19    HMP20 0.4183059           Stool         Female
#> 13    HMP18    HMP21 0.7170374 Saliva vs Stool           Male
#> 14    HMP19    HMP21 0.3896741           Stool Female vs Male
#> 15    HMP20    HMP21 0.1490926           Stool Female vs Male
    
    # Only look at distances among the stool samples
    bdiv_table(biom, 'unifrac', md = c("==Body Site", "Sex"))
#>    .sample1 .sample2 .distance Body Site            Sex
#> 10    HMP19    HMP20 0.4183059     Stool         Female
#> 14    HMP19    HMP21 0.3896741     Stool Female vs Male
#> 15    HMP20    HMP21 0.1490926     Stool Female vs Male
    
    # Or between males and females
    bdiv_table(biom, 'unifrac', md = c("Body Site", "!=Sex"))
#>    .sample1 .sample2 .distance       Body Site            Sex
#> 5     HMP18    HMP19 0.6651627 Saliva vs Stool Female vs Male
#> 9     HMP18    HMP20 0.6810017 Saliva vs Stool Female vs Male
#> 14    HMP19    HMP21 0.3896741           Stool Female vs Male
#> 15    HMP20    HMP21 0.1490926           Stool Female vs Male