Split BIOM by metadata, apply function, and return results in a data frame.

Split BIOM by metadata, apply function, and return results in a data frame.

bdply(biom, vars, FUN, ..., fast = TRUE)

Arguments

biom
A BIOM object, as returned from read.biom.

vars
A character vector of metadata fields. Each unique combination of values in these columns will be used to create a subsetted BIOM object to pass to FUN. If NULL, biom will be passed to FUN unaltered. Unambiguous abbreviations of metadata fields are also accepted.

FUN
The function to execute on each BIOM subset. FUN should return a data.frame, all of which will be rbind-ed together before being returned from bdply.

...
Additional arguments to pass on to FUN.

fast
If TRUE (the default), the subsetted BIOM objects will still contain the full taxa table and phylogenetic tree. Set fast = FALSE to run the slow steps of subsetting these elements as well.

Value

A data.frame comprising the merged outputs of FUN, along with the columns specified by vars.

Examples

    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read.biom(infile)
    
    bdply(biom, "Sex", nsamples)

#>      Sex V1
#> 1 Female 30
#> 2   Male 20

    
    bdply(biom, c("Body Site", "Sex"), function (b) {
      ad <- alpha.div(b)[,c("Shannon", "Simpson")]
      apply(ad, 2L, mean)
    })

#>        Body Site    Sex   Shannon   Simpson
#> 1 Anterior nares Female 1.4318071 0.6814534
#> 2 Anterior nares   Male 1.5053133 0.6650060
#> 3  Buccal mucosa Female 1.1739844 0.4080621
#> 4  Buccal mucosa   Male 1.7128815 0.6016869
#> 5     Mid vagina Female 0.4070774 0.1670592
#> 6         Saliva Female 2.9304968 0.8934531
#> 7         Saliva   Male 3.1718522 0.9132980
#> 8          Stool Female 2.4288307 0.8496172
#> 9          Stool   Male 2.5126148 0.8347696

    
    bdply(biom, "Body Site", function (b) {
      r <- range(beta.div(b, "bray"))
      data.frame(bray.min = r[[1]], bray.max = r[[2]])
    })

#>        Body Site   bray.min  bray.max
#> 1 Anterior nares 0.21357850 0.8957952
#> 2  Buccal mucosa 0.11132016 0.7878359
#> 3     Mid vagina 0.04589041 1.0000000
#> 4         Saliva 0.37105391 0.6975318
#> 5          Stool 0.40223744 0.9739696