Make a distance matrix of samples vs samples.

Make a distance matrix of samples vs samples.

beta.div(biom, method, weighted = TRUE, tree = NULL, progressbar = FALSE)

Arguments

biom
A matrix, simple_triplet_matrix, or BIOM object, as returned from read.biom. For matrices, the rows and columns are assumed to be the taxa and samples, respectively.

method
The distance algorithm to use. Options are: “manhattan”, “euclidean”, “bray-curtis”, “jaccard”, and “unifrac”. Non-ambiguous abbrevations of the method names are also accepted. A phylogentic tree must be present in biom or explicitly provided via tree= to use the UniFrac methods.

weighted
Take relative abundances into account. When weighted=FALSE, only presence/absence is considered.

tree
A phylo object representing the phylogenetic relationships of the taxa in biom. Will be taken from the tree embedded in the biom object if not explicitly specified. Only required for computing UniFrac distance matrices.

progressbar
Whether to display a progress bar and status messages (TRUE/FALSE). Will automatically tie in with shiny if run within a shiny session.

Value

A distance matrix.

Examples

    library(rbiom)

    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    biom <- select(biom, 1:10)

    dm <- beta.div(biom, 'unifrac')

#> Error in par_unifrac(counts, tree, ifelse(weighted, 1L, 0L)): object '_rbiom_par_unifrac' not found


    as.matrix(dm)[1:4,1:4]

#> Error in as.matrix(dm): object 'dm' not found

    plot(hclust(dm))

#> Error in hclust(dm): object 'dm' not found