Create a BIOM object.

Create a BIOM object.

Usage

biom_build(
  counts,
  metadata = NULL,
  taxonomy = NULL,
  tree = NULL,
  sequences = NULL,
  id = NULL,
  comment = NULL
)

Arguments

counts

The count data as a numeric matrix, where column names are the sample names and row names are the OTU names. A BIOM object or a filename/URL compatible with read_biom() is also acceptable.

metadata

A data.frame with sample names as the row names and metadata names as the column names. Or a filename/URL with comma- or tab-separated data.

taxonomy

A character matrix with OTU names as the row names. Or a filename/URL with comma- or tab-separated data.

tree

A phylo object with tip labels matching the OTU names. (E.g. from read_tree()). Or a filename/URL with newick formatted data.

sequences

A named character vector of DNA sequences. Or a filename/URL with fasta format data. (Currently not used by rbiom.)

id, comment

A character vector of length one with text of the user's choosing.

Value

A BIOM class object containing the parsed data. This object can be treated as a list with the following named elements:

$counts - A numeric slam sparse matrix of observation counts. Taxa (OTUs) as rows and samples as columns.

$metadata - A data frame containing any embedded metadata. Row names are sample IDs.

$taxonomy - Character matrix of taxonomic names, if given. Row names are taxa (OTU) IDs. Column rows are named Kingdom, Phylum, Class, Order, Family, Genus, Species, and Strain, or TaxLvl.1, TaxLvl.2, ... , TaxLvl.N when more than 8 levels of taxonomy are encoded in the biom file.

$phylogeny - An object of class phylo defining the phylogenetic relationships between the taxa. Although the official specification for BIOM only includes phylogenetic trees in BIOM version 2.1, if a BIOM version 1.0 file includes a phylogeny entry with newick data, then it will be loaded here as well. The ape package has additional functions for working with phylo objects.

$sequences - A named character vector, where the names are taxonomic identifiers and the values are the sequences they represent. These values are not part of the official BIOM specification, but will be read and written when defined.

$info - A list of other attributes defined in the BIOM file, such as id, type, format, format_url, generated_by, date, matrix_type, matrix_element_type, Comment, and shape.

metadata, taxonomy, and phylogeny are optional components of the BIOM file specification and therefore will be empty in the returned object when they are not provided by the BIOM file.

Examples

    library(rbiom)
    
    biom <- sample_select(hmp50, 1:5)
    
    # Save components to separate files.
    ct  <- write_counts(biom, tempfile())
    md  <- write_metadata(biom, tempfile())
    tax <- write_taxonomy(biom, tempfile())
    tre <- write_tree(biom, tempfile())
    fas <- write_fasta(biom, tempfile())
    
    # Peek at the file structures.
    cat(readLines(ct, n = 2L),  '', sep="\n")
#> #OTU ID	HMP01	HMP02	HMP03	HMP04	HMP05
#> UncO4101	1	5	6	18	5
#> 
    cat(readLines(md, n = 2L),  '', sep="\n")
#> SampleID	Age	BMI	Body Site	Sex 
#> HMP01	22	20	Buccal mucosa	Female
#> 
    cat(readLines(tax, n = 2L), '', sep="\n")
#> UncO4101	Bacteria	Firmicutes	Clostridia	Clostridiales	Lachnospiraceae	Lachnoanaerobaculum
#> Unc96922	Bacteria	Bacteroidetes	Flavobacteriia	Flavobacteriales	Flavobacteriaceae	Capnocytophaga
#> 
    cat(readChar(tre, nchars = 50L), "\n\n")
#> ((((((((((Unc92490:0.05748,Unc02pjp:0.01233):0.007 
#> 
    cat(readChar(fas, nchars = 50L), "\n\n")
#> >UncO4101
#> TGGGGAATATTGGACAATGGGGGAAACCCTGATCCAGCG 
#> 
    
    # Re-assemble the BIOM object.
    biom <- biom_build(ct, md, tax, tre, id = "New BIOM")
    print(biom)
#> New BIOM (2023-10-03)
#>       5 Samples:  (HMP01, HMP02, HMP03, HMP04, and HMP05)
#>     132 Taxa:     (UncO4101, Unc96922, Unc02k4w, Unc1721...)
#>       6 Ranks:    (Kingdom, Phylum, Class, Order, Family...)
#>       4 Metadata: (Age, BMI, Body Site, and Sex)
#>         Tree:     Present
#> 
    
    # Remove temporary files.
    invisible(file.remove(c(ct, md, tax, fas, tre)))