Combine several rbiom objects into one.

WARNING: It is generally ill-advised to merge BIOM datasets, as OTUs mappings are dependent on upstream clustering and are not equivalent between BIOM files.

Usage

biom_merge(...)

Arguments

...

Any number of rbiom objects (e.g. from read_biom()), lists of rbiom objects, or valid arguments to the src parameter of read_biom() (for instance file names).

Value

An rbiom class object containing the parsed data. This object can be treated as a list with the following named elements:

$counts - A numeric slam::simple_triplet_matrix() (sparse matrix) of observation counts. Taxa (OTUs) as rows and samples as columns. Access or modify using otu_matrix().

$metadata - A tibble::tibble() (data frame) containing any embedded metadata. Sample IDs are in the .sample column. Access or modify using sample_metadata().

$taxonomy - A tibble::tibble() (data frame) mapping OTU IDs to taxonomic clades. Columns are named .otu, Kingdom, Phylum, Class, Order, Family, Genus, Species, and Strain, or TaxLvl.1, TaxLvl.2, ... , TaxLvl.N when more than 8 levels of taxonomy are encoded in the biom file. Access or modify using otu_taxonomy().

$phylogeny - An object of class phylo defining the phylogenetic relationships between the OTUs. Although the official specification for BIOM only includes phylogenetic trees in BIOM version 2.1, if an rbiom version 1.0 file includes a phylogeny entry with newick data, then it will be loaded here as well. The ape package has additional functions for working with phylo objects. Access or modify using otu_tree().

$sequences - A named character vector, where the names are OTU IDs and the values are the nucleic acid sequences they represent. These values are not part of the official BIOM specification, but will be read and written when defined. Access or modify using otu_sequences().

$info - A list of other attributes defined in the BIOM file, such as id, type, format, format_url, generated_by, date, matrix_type, matrix_element_type, comment, and shape. Access using biom_info(). Access/set with biom_id() or biom_comment().

metadata, taxonomy, and phylogeny are optional components of the BIOM file specification and therefore will be NULL or simple placeholders in the returned object when they are not provided by the BIOM file.

Examples

    library(rbiom)
    
    b1 <- sample_select(hmp50, 1:4)
    b2 <- sample_select(hmp50, 5:8)
    
    biom <- biom_merge(b1, b2)
    print(biom)
#> Merged BIOM (2023-11-22)
#>       8 Samples:  (HMP01, HMP02, HMP03, HMP04, HMP05, HM...)
#>     168 Taxa:     (UncO4101, Unc96922, Unc02k4w, Unc1721...)
#>       7 Ranks:    (.otu, Kingdom, Phylum, Class, Order, ...)
#>       5 Metadata: (.sample, Age, BMI, Body Site, and Sex)
#>         Tree:     Absent
#> 
    
    otu_tree(biom) <- otu_tree(hmp50)
    print(biom)
#> Merged BIOM (2023-11-22)
#>       8 Samples:  (HMP01, HMP02, HMP03, HMP04, HMP05, HM...)
#>     168 Taxa:     (UncO4101, Unc96922, Unc02k4w, Unc1721...)
#>       7 Ranks:    (.otu, Kingdom, Phylum, Class, Order, ...)
#>       5 Metadata: (.sample, Age, BMI, Body Site, and Sex)
#>         Tree:     Present
#>