Split BIOM by metadata, apply function, and return results in a list.

Calls plyr::dlply internally.

blply(biom, vars, FUN, ..., fast = TRUE)

Arguments

biom

A BIOM object, as returned from read_biom.

vars

A character vector of metadata fields. Each unique combination of values in these columns will be used to create a subsetted BIOM object to pass to FUN. Unambiguous abbreviations of metadata fields are also accepted.

FUN

The function to execute on each BIOM subset. FUN may return any object, all of which will be returned in a named list.

...

Additional arguments to pass on to FUN.

fast

If TRUE (the default), the subsetted BIOM objects will still contain the full taxa table and phylogenetic tree. Set fast = FALSE to run the slow steps of subsetting these elements as well.

Value

A list with the function outputs.

Examples

    library(rbiom)
    
    blply(hmp50, "Sex", nsamples)
#> $Female
#> [1] 30
#> 
#> $Male
#> [1] 20
#> 
#> attr(,"split_type")
#> [1] "data.frame"
#> attr(,"split_labels")
#>      Sex
#> 1 Female
#> 2   Male
    
    blply(hmp50, c("Body Site", "Sex"), function (b) {
      ad <- adiv_table(b)[,c("Shannon", "Simpson")]
      apply(ad, 2L, mean)
    })
#> $`Female.Anterior nares`
#>   Shannon   Simpson 
#> 1.4318071 0.6814534 
#> 
#> $`Female.Buccal mucosa`
#>   Shannon   Simpson 
#> 1.1739844 0.4080621 
#> 
#> $`Female.Mid vagina`
#>   Shannon   Simpson 
#> 0.4070774 0.1670592 
#> 
#> $Female.Saliva
#>   Shannon   Simpson 
#> 2.9304968 0.8934531 
#> 
#> $Female.Stool
#>   Shannon   Simpson 
#> 2.4288307 0.8496172 
#> 
#> $`Male.Anterior nares`
#>  Shannon  Simpson 
#> 1.505313 0.665006 
#> 
#> $`Male.Buccal mucosa`
#>   Shannon   Simpson 
#> 1.7128815 0.6016869 
#> 
#> $Male.Saliva
#>  Shannon  Simpson 
#> 3.171852 0.913298 
#> 
#> $Male.Stool
#>   Shannon   Simpson 
#> 2.5126148 0.8347696 
#> 
#> attr(,"split_type")
#> [1] "data.frame"
#> attr(,"split_labels")
#>      Sex      Body Site
#> 1 Female Anterior nares
#> 2 Female  Buccal mucosa
#> 3 Female     Mid vagina
#> 4 Female         Saliva
#> 5 Female          Stool
#> 6   Male Anterior nares
#> 7   Male  Buccal mucosa
#> 8   Male         Saliva
#> 9   Male          Stool
    
    blply(hmp50, "Body Site", function (b) {
      r <- range(bdiv_distmat(b, "bray"))
      data.frame(bray.min = r[[1]], bray.max = r[[2]])
    })
#> $`Anterior nares`
#>    bray.min  bray.max
#> 1 0.2135785 0.8957952
#> 
#> $`Buccal mucosa`
#>    bray.min  bray.max
#> 1 0.1113202 0.7878359
#> 
#> $`Mid vagina`
#>     bray.min bray.max
#> 1 0.04589041        1
#> 
#> $Saliva
#>    bray.min  bray.max
#> 1 0.3710539 0.6975318
#> 
#> $Stool
#>    bray.min  bray.max
#> 1 0.4022374 0.9739696
#> 
#> attr(,"split_type")
#> [1] "data.frame"
#> attr(,"split_labels")
#>        Body Site
#> 1 Anterior nares
#> 2  Buccal mucosa
#> 3     Mid vagina
#> 4         Saliva
#> 5          Stool