Create a BIOM object.

Create a BIOM object.

Usage

build_biom(
  counts,
  metadata = NULL,
  taxonomy = NULL,
  phylogeny = NULL,
  sequences = NULL,
  id = NULL,
  comments = NULL
)

Arguments

counts

The count data as a numeric matrix, where column names are the sample names and row names are the taxa IDs. A BIOM object or a filename/URL compatible with read_biom() is also acceptable.

metadata

A data.frame with sample IDs as the row names and metadata names as the column names. Or a filename/URL with comma- or tab-separated data.

taxonomy

A character matrix with taxa IDs as the row names. Or a filename/URL with comma- or tab-separated data.

phylogeny

A phylo object with tip labels matching the taxa IDs. (E.g. from rbiom::read_tree()). Or a filename/URL with newick formatted data.

sequences

A named character vector of DNA sequences. Or a filename/URL with fasta format data. (Currently not used by rbiom.)

id, comments

A character vector of length one with text of the user's choosing.

Examples

    library(rbiom)
    
    biom <- select(hmp50, 1:5)
    
    # Save components to separate files.
    ct  <- write_counts(biom, tempfile())
    md  <- write_metadata(biom, tempfile())
    tax <- write_taxonomy(biom, tempfile())
    tre <- write_tree(biom, tempfile())
    fas <- write_fasta(biom, tempfile())
    
    # Peek at the file structures.
    cat(readLines(ct, n = 2L),  '', sep="\n")
#> #OTU ID	HMP01	HMP02	HMP03	HMP04	HMP05
#> UncO4101	1	5	6	18	5
#> 
    cat(readLines(md, n = 2L),  '', sep="\n")
#> SampleID	Age	BMI	Body Site	Experiment Center	Sequencing Platform	Sex 
#> HMP01	22	20	Buccal mucosa	BCM	FLX	Female
#> 
    cat(readLines(tax, n = 2L), '', sep="\n")
#> UncO4101	Bacteria	Firmicutes	Clostridia	Clostridiales	Lachnospiraceae	Lachnoanaerobaculum
#> Unc96922	Bacteria	Bacteroidetes	Flavobacteriia	Flavobacteriales	Flavobacteriaceae	Capnocytophaga
#> 
    cat(readChar(tre, nchars = 50L), "\n\n")
#> ((((((((((Unc92490:0.05748,Unc02pjp:0.01233):0.007 
#> 
    cat(readChar(fas, nchars = 50L), "\n\n")
#> >UncO4101
#> TGGGGAATATTGGACAATGGGGGAAACCCTGATCCAGCG 
#> 
    
    # Re-assemble the BIOM object.
    biom <- build_biom(ct, md, tax, tre, id = "New BIOM")
    print(biom)
#> New BIOM (2023-06-21)
#>       5 Samples:  (HMP01, HMP02, HMP03, HMP04, and HMP05)
#>     132 Taxa:     (UncO4101, Unc96922, Unc02k4w, Unc1721...)
#>       6 Ranks:    (Kingdom, Phylum, Class, Order, Family...)
#>       6 Metadata: (Age, BMI, Body Site, Experiment Cente...)
#>         Tree:     Present
#> 
    
    # Remove temporary files.
    invisible(file.remove(c(ct, md, tax, fas, tre)))