Run statistics on a distance matrix vs a categorical or numeric variable.

Run statistics on a distance matrix vs a categorical or numeric variable.

Usage

distmat_stats(
  dm,
  groups,
  test = "adonis2",
  seed = 0,
  permutations = 999,
  p.adj = "fdr"
)

Arguments

dm

The distance matrix as a dist object, or a list of dist objects.

groups

A named vector of grouping values. The names should correspond to attr(dm, 'Labels').

test

The name of the function from the vegan R package to use. Options are: "adonis2" and "mrpp". Non-ambiguous abbreviations of the function names are also accepted. Default: "adonis2"

seed

Random seed for permutations. Default: 0

permutations

Number of random permutations to use. Default: 999

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: "fdr".

Value

A data.frame with summary statistics from vegan::permustats(). The columns are:

stat - The observed statistic. For mrpp, this is the overall weighted mean of group mean distances.

z - The difference of observed statistic and mean of permutations divided by the standard deviation of permutations (also known as z-values). Evaluated from permuted values without observed statistic.

p.val - Include only the specified metadata columns.

adj.p - P-value after correcting for the false discovery rate with the p.adj method.

test - The test and p.adj methods used.

See also

Examples

    library(rbiom)
    
    biom <- sample_select(hmp50, 1:10)
    dm   <- bdiv_distmat(biom, 'unifrac')
    distmat_stats(dm, groups = sample_metadata(biom, 'Body Site'))
#>   .stat .z .p.val .adj.p        .test
#> 1  9.27 11  0.001  0.001 adonis2; fdr
    distmat_stats(dm, groups = sample_metadata(biom, 'Age'))
#>   .stat   .z .p.val .adj.p        .test
#> 1  4.13 3.56  0.012  0.012 adonis2; fdr