documentation_corrplot

documentation_corrplot

Arguments

x

A numeric metadata field to use for the x-axis. Required.

layers

"trend", "scatter". Single letter abbreviations are also accepted. For instance, c("trend", "scatter") is equivalent to c("t", "s") and "ts". See vignette("corrplots") for examples of each. Default: "t".

color.by, facet.by, limit.by

Metadata columns to use for aesthetics and partitioning. See below for details. Default: NULL

test

Which test statistic to display on the plot. Options are:

"fit" - How well does the model fit the data?

"terms" - How strongly does 'x' influence 'y'?

"means" - Is the average 'y' value non-zero?

"trends" - Does any trendline have a non-zero slope?

"pw_means" - Are the means of any trendlines different?

"pw_trends" - Are the slopes of any trendlines different?

"none" - Do not compute or show statistics.

Default: "trends"

Note: "pw_means" and "pw_trends" can only be calculated when using a color.by metadata column with more than one level.

Statistical tests are run separately on each facet. P-values are adjusted for multiple comparisons by considering all facets together.

...

Additional parameters to pass along to ggplot2 functions. Prefix a parameter name with either t. or s./pt. to ensure it gets passed to (and only to) geom_smooth or geom_point, respectively. For instance, s.size = 2 ensures only the scatterplot points have their size set to 2.

Value

A ggplot2 plot.
The computed data points, statistics, and ggplot command are available as $data, $stats, and $code, respectively.

Aesthetics and Partitions

Metadata can be used to flexibly subset, partition, and apply aesthetics when creating a plot. Common use cases are provided below. More thorough documentation is available at https://cmmr.github.io/rbiom.

  ## Colors ----------------------------
  color.by = "Body Site"
  color.by = list('Body Site' = "bright")
  color.by = list('Body Site' = c("Stool", "Saliva"))
  color.by = list('Body Site' = list('values' = c("Stool", "Saliva"), 'colors' = "bright"))
  color.by = list('Body Site' = c('Stool' = "blue", 'Saliva' = "green"))
  
  ## Facets ----------------------------
  facet.by = "Body Site"
  facet.by = c("Body Site", "Sex")
  facet.by = list('Body Site' = c("Stool", "Saliva"), "Sex")
  
  ## Limits ----------------------------
  limit.by = list('Sex' = "Male", 'Age' = c(20,40))
  limit.by = list('Body Site' = c("Saliva", "Anterior nares"), 'Age' = c(NA,35))

color.by - Any metadata column. (Max 1)

facet.by - Only categorical metadata column(s).

limit.by - Any metadata column(s).

All built-in color palettes are colorblind-friendly.

The available categorical palette names are: "okabe", "carto", "r4", "polychrome", "tol", "bright", "light", "muted", "vibrant", "tableau", "classic", "alphabet", "tableau20", "kelly", and "fishy".

The available numeric palette names are: "reds", "oranges", "greens", "purples", "grays", "acton", "bamako", "batlow", "bilbao", "buda", "davos", "devon", "grayC", "hawaii", "imola", "lajolla", "lapaz", "nuuk", "oslo", "tokyo", "turku", "bam", "berlin", "broc", "cork", "lisbon", "roma", "tofino", "vanimo", "vik"