Get or set the taxonomy table.

Get or set the taxonomy table.

Usage

otu_taxonomy(biom, rank = NULL, unc = "asis", lineage = FALSE)

otu_taxonomy(biom) <- value

Arguments

biom

An rbiom-class object, or data coercible with as_rbiom().

rank

A single taxonomic rank to return as a named vector, or NULL to return the complete table. Default: NULL

unc

How to handle unclassified, uncultured, and similarly ambiguous taxa names. Options are:

"singly" - Replaces them with the OTU name.

"grouped" - Replaces them with a higher rank's name.

"drop" - Excludes them from the result.

"asis" - To not check/modify any taxa names.

Default: "singly"

Abbreviations are allowed.

lineage

Include all ranks in the name of the taxa. For instance, setting to TRUE will produce Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales. Otherwise the taxa name will simply be Coriobacteriales. You want to set this to TRUE when unc = "asis" and you have taxa names (such as Incertae_Sedis) that map to multiple higher level ranks. Default: FALSE

value

An object coercible with tibble::as_tibble(). Must either have rownames or an '.otu' column with OTU names matching otu_names(x).

Value

Depending on rank, a named character vector or a tibble data frame with '.otu' as the first column name.

See also

Examples

    library(rbiom) 
    
    # Display the full taxonomic data for each OTU ------------------------
    otu_taxonomy(hmp50)
#> # A tibble: 490 × 7
#>    Kingdom  Phylum           Class              Order         Family Genus .otu 
#>    <fct>    <fct>            <fct>              <fct>         <fct>  <fct> <fct>
#>  1 Bacteria Bacteroidetes    Bacteroidia        Bacteroidales Riken… Alis… UncO…
#>  2 Bacteria Firmicutes       Clostridia         Clostridiales Lachn… Lach… UncO…
#>  3 Bacteria Actinobacteria   Actinobacteria     Actinomyceta… Actin… Acti… AnmM…
#>  4 Bacteria Bacteroidetes    Bacteroidia        Bacteroidales Prevo… Prev… PreB…
#>  5 Bacteria Bacteroidetes    Bacteroidia        Bacteroidales Porph… Barn… CprS…
#>  6 Bacteria Bacteroidetes    Flavobacteriia     Flavobacteri… Flavo… Capn… Unc9…
#>  7 Bacteria Proteobacteria   Betaproteobacteria Burkholderia… Comam… Coma… Unc0…
#>  8 Bacteria Firmicutes       Clostridia         Clostridiales Pepto… Pept… Unc0…
#>  9 Bacteria Saccharibacteria Unknown Class      Unknown Order Unkno… Cand… Unc1…
#> 10 Bacteria Firmicutes       Bacilli            Bacillales    Liste… Broc… BroS…
#> # ℹ 480 more rows
    
    # Only show the OTU -> Genus mapping  ---------------------------------
    otu_taxonomy(hmp50, "Genus") %>% head()
#>            UncO2713            UncO4101            AnmMass2            PreBivi6 
#>           Alistipes Lachnoanaerobaculum         Actinomyces          Prevotella 
#>            CprSpeci            Unc96922 
#>         Barnesiella      Capnocytophaga 
#> 201 Levels: Abiotrophia Acidaminococcus Acinetobacter ... Veillonella
    
    # Sometimes taxonomic names are incomplete ----------------------------
    otu_taxonomy(hmp50)[,4:7] %>%
      dplyr::filter(.otu %in% c('GemAsacc', 'GcbBacte', 'Unc58411'))
#> # A tibble: 3 × 4
#>   Order          Family    Genus   .otu    
#>   <fct>          <fct>     <fct>   <fct>   
#> 1 Bacillales     Family XI Gemella GemAsacc
#> 2 o              f         g       GcbBacte
#> 3 Mollicutes RF9 f         g       Unc58411
    
    # rbiom can insert more descriptive placeholders ----------------------
    otu_taxonomy(hmp50, unc = "singly")[,4:7] %>%
      dplyr::filter(.otu %in% c('GemAsacc', 'GcbBacte', 'Unc58411'))
#> # A tibble: 3 × 4
#>   Order          Family        Genus         .otu    
#>   <fct>          <fct>         <fct>         <fct>   
#> 1 Bacillales     Bacillales XI Gemella       GemAsacc
#> 2 Unc. GcbBacte  Unc. GcbBacte Unc. GcbBacte GcbBacte
#> 3 Mollicutes RF9 Unc. Unc58411 Unc. Unc58411 Unc58411
    
    # Or collapse them into groups ----------------------------------------
    otu_taxonomy(hmp50, unc = "grouped")[,4:7] %>%
      dplyr::filter(.otu %in% c('GemAsacc', 'GcbBacte', 'Unc58411'))
#> # A tibble: 3 × 4
#>   Order                Family               Genus                .otu    
#>   <fct>                <fct>                <fct>                <fct>   
#> 1 Bacillales           Bacillales XI        Gemella              GemAsacc
#> 2 Unc. Gracilibacteria Unc. Gracilibacteria Unc. Gracilibacteria GcbBacte
#> 3 Mollicutes RF9       Unc. Mollicutes RF9  Unc. Mollicutes RF9  Unc58411