Deprecated functions in package rbiom.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("<function>-deprecated").

Usage

alpha.div(biom, rarefy = FALSE)

beta.div(
  biom,
  method = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  long = FALSE,
  md = FALSE
)

counts(biom)

info(biom)

metadata(biom, field = NULL, cleanup = FALSE)

nsamples(biom)

ntaxa(biom)

phylogeny(biom)

read.biom(src, tree = "auto", prune = FALSE)

read.fasta(file, ids = NULL)

read.tree(src)

sample.names(biom)

# S3 method for class 'rbiom'
select(
  .data,
  samples = NULL,
  nTop = NULL,
  nRandom = NULL,
  seed = 0,
  biom = NULL,
  ...
)

sequences(biom)

subtree(tree, tips)

taxa.names(biom)

taxa.ranks(biom)

taxa.rollup(
  biom,
  rank = "OTU",
  map = NULL,
  lineage = FALSE,
  sparse = FALSE,
  taxa = NULL,
  long = FALSE,
  md = FALSE
)

taxonomy(biom, ranks = NULL, unc = "asis")

tips(x)

unifrac(biom, weighted = TRUE, tree = NULL)

write.biom(biom, file, format = "json")

write.fasta(seqs, outfile = NULL)

write.tree(tree, file = NULL)

write.xlsx(biom, outfile, depth = 0.1, seed = 0)

as.percent(biom)

comments(biom)

depth(biom)

depths_barplot(
  biom,
  rline = TRUE,
  counts = TRUE,
  labels = TRUE,
  transform = "log10",
  ...
)

has.phylogeny(biom)

has.sequences(biom)

id(biom)

is.rarefied(biom)

repair(biom)

sample_subset(x, ...)

sample.sums(biom, long = FALSE, md = FALSE)

taxa_max(biom, rank = -1, lineage = FALSE, unc = "singly")

taxa.means(biom, rank = NULL)

taxa.sums(biom, rank = NULL)

top.taxa(biom, rank = "OTU", n = Inf)

top_taxa(biom, rank = "OTU", n = Inf)

comments(x) <- value

counts(x) <- value

id(x) <- value

metadata(x) <- value

phylogeny(x) <- value

sample.names(x) <- value

sequences(x) <- value

taxa.names(x) <- value

taxa.ranks(x) <- value

taxonomy(x) <- value

alpha.div

Use adiv_matrix() or adiv_table() instead.

beta.div

Use bdiv_table() or bdiv_distmat() instead.

counts

Use $counts instead.

info

Use biom$id, biom$comment, etc instead.

metadata

Use biom$metadata or pull(biom, field) instead.

nsamples

Use biom$n_samples instead.

ntaxa

Use biom$n_otus instead.

phylogeny

Use biom$tree instead.

read.biom

Use as_rbiom() instead.

read.fasta

Use read_fasta() instead.

read.tree

Use read_tree() instead.

sample.names

Use biom$samples instead.

select

Use slice() instead.

sequences

Use biom$sequences instead.

subtree

Use tree_subset() instead.

taxa.names

Use biom$otus instead.

taxa.ranks

Use biom$ranks instead.

taxa.rollup

Use taxa_table() taxa_matrix() instead.

taxonomy

Use $taxonomy instead.

tips

Use tree$tip.label instead.

unifrac

Use bdiv_distmat() or bdiv_table() instead.

write.biom

Use write_biom() instead.

write.fasta

Use write_fasta() instead.

write.tree

Use write_tree() instead.

write.xlsx

Use write_xlsx() instead.

as.percent

Use biom$counts %<>% rescale_cols() instead.

comments

Use biom$comment instead.

depth

Use sample_sums() instead.

depths_barplot

Use rare_stacked() instead.

has.phylogeny

Use !is.null(biom$tree) instead.

has.sequences

Use !is.null(biom$sequences) instead.

id

Use biom$id instead.

is.rarefied

Use !is.null(biom$depth) instead.

repair

Use as_rbiom(as.list(biom)) instead.

sample_subset

Use biom$metadata %<>% base::subset() instead.

sample.sums

Use sample_sums() or adiv_table() instead.

taxa_max

Use taxa_apply(biom, max, sort = 'desc') instead.

taxa.means

Use taxa_means() instead.

taxa.sums

Use taxa_sums() instead.

top.taxa

Use taxa_sums() instead.

top_taxa

Use taxa_sums() instead.

comments-set

Use biom$comment <- instead.

counts-set

Use biom$counts <- instead.

id-set

Use biom$id <- instead.

metadata-set

Use biom$metadata <- instead.

phylogeny-set

Use biom$tree <- instead.

sample.names-set

Use biom$samples <- instead.

sequences-set

Use biom$sequences <- instead.

taxa.names-set

Use biom$otus <- instead.

taxa.ranks-set

Use biom$ranks <- instead.

taxonomy-set

Use biom$taxonomy <- instead.