Working with rbiom Objects.

Rbiom objects make it easy to access and manipulate your BIOM data, ensuring all the disparate components remain in sync. These objects behave largely like lists, in that you can access and assign to them using the $ operator. The sections below list all the fields which can be read and/or written, and the helper functions for common tasks like rarefying and subsetting. To create an rbiom object, see as_rbiom().

Readable Fields

Reading from fields will not change the rbiom object.

Accessor	Content

$counts	Abundance of each OTU in each sample.

$metadata	Sample mappings to metadata (treatment, patient, etc).

$taxonomy	OTU mappings to taxonomic ranks (genus, phylum, etc).

$otus, $n_otus	OTU names.

$samples, $n_samples	Sample names.

$fields, $n_fields	Metadata field names.

$ranks, $n_ranks	Taxonomic rank names.

$tree, $sequences	Phylogenetic tree / sequences for the OTUs, or NULL.

$id, $comment	Arbitrary strings for describing the dataset, or NULL.

$depth	Rarefaction depth, or NULL if unrarefied.

Writable Fields

Assigning new values to these components will trigger validation checks and inter-component synchronization.

Component	What can be assigned.

$counts	matrix of abundances; OTUs (rows) by samples (columns)

$metadata	data.frame with '.sample' as the first column

$taxonomy	data.frame with '.otu' as the first column

$tree	phylo object with the phylogenetic tree for OTUs

$sequences	character vector of OTU reference sequences

$id, $comment	string with a title or comment for the dataset

Transformations

Function	Transformation

$clone()	Safely duplicate an rbiom object.

subset()	Subset samples according to metadata properties.

select()	Keep just the specified sample names or indices.

rarefy()	Sub-sample OTU counts to an even sampling depth.

All functions return an rbiom object.

Clone an Object

Use $clone() to create a copy of an rbiom object. This is necessary because rbiom objects are passed by reference. The usual <- assignment operator will simply create a second reference to the same object - it will not create a second object.

Usage

$clone(deep = FALSE)

Arguments

deep Not used.

Examples

biom <- hmp50$clone()