Transform a counts matrix.

Rarefaction subset counts so that all samples have the same number of observations. Rescaling rows or cols scales the matrix values so that row sums or column sums equal 1.

Usage

rarefy_cols(mtx, depth = 0.1, n = NULL, seed = 0)

rescale_cols(mtx)

rescale_rows(mtx)

Arguments

mtx

A matrix-like object.

depth

The number of observations to keep, per column (sample). If set to 'auto', a depth will be automatically selected. Samples that have fewer than this number of observations will be dropped. If called on data with non-integer abundances, values will be re-scaled to integers between 1 and depth such that they sum to depth. Default: 'auto'

n

The number of samples to retain, or if negative, the number of samples to drop. If NULL, the number kept/dropped is automatically selected. When not NULL, overrides any value given to depth. Default: NULL

seed

An integer to use for seeding the random number generator. If you need to create different random rarefactions of the same matrix, set this seed value to a different number each time.

Value

The rarefied or rescaled matrix.

See also

Examples

    library(rbiom)
    
    # rarefy_cols --------------------------------------
    biom <- hmp50$clone()
    sample_sums(biom) %>% head(10)
#> HMP01 HMP02 HMP03 HMP04 HMP05 HMP06 HMP07 HMP08 HMP09 HMP10 
#>  1660  1371  1353  1895  3939  4150  3283  1695  2069  2509 

    biom$counts %<>% rarefy_cols(depth=1000)
    sample_sums(biom) %>% head(10)
#> HMP01 HMP02 HMP03 HMP04 HMP05 HMP06 HMP07 HMP08 HMP09 HMP10 
#>  1000  1000  1000  1000  1000  1000  1000  1000  1000  1000 
    
    
    # rescaling ----------------------------------------
    mtx <- matrix(sample(1:20), nrow=4)
    mtx
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]   16    2    6   12   11
#> [2,]   17    4    3    1   10
#> [3,]   18   13    9    8   20
#> [4,]   19    5   15    7   14
    
    rowSums(mtx)
#> [1] 47 35 68 60
    rowSums(rescale_rows(mtx))
#> [1] 1 1 1 1
    
    colSums(mtx)
#> [1] 70 24 33 28 55
    colSums(rescale_cols(mtx))
#> [1] 1 1 1 1 1