Reduce samples to a specific list

Reduce samples to a specific list

select(
  biom,
  samples = NULL,
  nTop = NULL,
  nRandom = NULL,
  seed = 0,
  fast = FALSE
)

Arguments

biom

A BIOM object, as returned from read_biom. Objects of class matrix or simple_triplet_matrix are also accepted.

samples

Sample names, indices, or a logical vector identifying the samples to keep. The latter two should be based on the order of sample names given by colnames(biom$counts).

nTop

Selects this number of samples, taking the sample with the most observations first, then the sample with the second-most observations, etc. Ties will be randomly ordered. If nTop is higher than the number of samples in the dataset, the entire dataset will be returned. See note.

nRandom

Randomly selects this number of samples. If higher than the number of samples in the dataset, the entire dataset will be returned. See note.

seed

Random seed, used when selecting nRandom samples.

fast

Should subsetting the phylogenetic tree and sequences be skipped? These slow steps are often not necessary. (Default: FALSE)

Note: Generally, you will specify only one of the filters: samples, nTop, or nRandom. However, specifying multiple filters is allowed; they will be applied in the order listed above.

Value

An object of the same class as biom, subsetted as requested.

See also

Examples

    library(rbiom)
    
    ex1 <- select(hmp50, c('HMP14', 'HMP22', 'HMP03'))
    ex2 <- select(hmp50, c(32, 11, 28, 16, 46, 5))
    ex3 <- select(hmp50, 1:50 %% 6 == 0)
    ex4 <- select(hmp50, nRandom = 10)
    ex5 <- select(hmp50, nTop = 5)
    ex6 <- select(hmp50, samples = 10:40, nTop = 20, nRandom = 10)