Compute p-values for data vs metadata.

Compute p-values for data vs metadata.

stats.table(
  biom,
  x,
  y,
  by = NULL,
  adj = "fdr",
  pairwise = FALSE,
  weighted = TRUE,
  digits = 3,
  y.pos = NULL,
  y.pos.facet = "Metric"
)

Arguments

biom
A BIOM object, as returned from read.biom. Alternatively, a data.frame with column names expected by x, y, and by.

x
The metadata column name of the category to use for comparisons.

y
A character vector with values from one of the following sets:

Alpha Diversity Metrics (one or more)

OTUs, Shannon, Chao1, Simpson, and/or InvSimpson. You may also set y = "Diversity" to get all five metrics.

Beta Diversity Metrics (one only)

manhattan, euclidean, bray-curtis, jaccard, or unifrac. Use in combination with the weighted parameter. Metadata column names can be prefixed with == or != to limit distance calculations to within or between, respectively, those categories. See examples below. Setting y = "Distance" will use unifrac if a phylogenetic tree is present, or bray-curtis otherwise.

Taxa Abundances (one only)

Kingdom, Phylum, Class, Order, Family, Genus, Species, Strain, or OTU. Supported ranks will vary by biom. Run taxa.ranks(biom) to see the available options. Specifying y = "Taxa" will default to the most precise rank possible.

by
Additional metadata columns to group by (e.g. facets in a plot).

adj
Adjustment to use for multiple comparisons. See stats::p.adjust.methods for available options. Default: fdr.

pairwise
If FALSE (the default), one p-value is generated for each by group, or 1 p-value when by = NULL. If TRUE, unique values in x are compared pairwise against each other.

weighted
If TRUE(the default), run beta diversity metrics in weighted mode. Set to FALSE to use unweighted mode.

digits
Round p-values to this many significant figures (default 3). Set to NULL to disable rounding.

y.pos
Add an extra column to the results to indicate the upper most value of each group. Added for ggpubr::geom_bracket(). Options are NULL for no y.pos column (the default), max for the maximum value, box for a box plot's the whisker upper bound, or violin for the highest point of a violin plot.

Value

A data.frame with columns p.val, adj.p, adj.sig, as well as columns for tracking for x, y, and by categories.

Examples

    library(rbiom)
    
    biom <- rarefy(hmp50)
    biom <- subset(biom, `Body Site` %in% c('Saliva', 'Stool', 'Buccal mucosa'))
      
    
    # Overall, do the three body sites different have different Shannon
    #  diversity indices (alpha diversity)? - YES
    stats.table(biom, x = "Body Site", y = "Shannon")

#>    n           Test    p.val    adj.p
#> 1 29 Kruskal-Wallis 4.68e-05 4.68e-05

    
    # Which body sites have the most different Shannon diversity
    #  indices? - Buccal mucosa vs. Saliva
    stats.table(biom, x = "Body Site", y = "Shannon", pairwise = TRUE)

#>    n         Test Group1        Group2    p.val    adj.p
#> 1 20 Mann-Whitney Saliva Buccal mucosa 2.17e-05 0.000065
#> 2 19 Mann-Whitney  Stool Buccal mucosa 1.45e-03 0.002180
#> 3 19 Mann-Whitney Saliva         Stool 2.99e-03 0.002990

    
    # Do males or females have more alpha diversity variation among 
    #  body sites? - Females
    stats.table(biom, x = "Body Site", y = "Shannon", by = "Sex")

#>      Sex  n           Test   p.val   adj.p
#> 1 Female 15 Kruskal-Wallis 0.00374 0.00747
#> 2   Male 14 Kruskal-Wallis 0.00821 0.00821

    
    # Which phylum is most different overall between males and females?
    #  - Saccharibacteria
    stats.table(biom, x = "Sex", y = "Phylum") %>% head()

#>     n         Test p.val adj.p
#> 1 348 Mann-Whitney 0.242 0.242

    
    # Which phyla is the most differentially abundant between males and 
    #  females on a particular body site? - Tenericutes in Saliva
    stats.table(biom, x = "Sex", y = "Phylum", by = "Body Site") %>% head()

#>       Body Site   n         Test p.val adj.p
#> 1        Saliva 120 Mann-Whitney 0.232 0.695
#> 2 Buccal mucosa 120 Mann-Whitney 0.819 0.909
#> 3         Stool 108 Mann-Whitney 0.909 0.909

    
    # Overall, are males and females characterized by their OTU 
    #  abundances? - NO
    stats.table(biom, x = "Sex", y = "unifrac")

#>     n           Test p.val adj.p
#> 1 406 Kruskal-Wallis  0.49  0.49

    
    # What about on a particular body site? - Yes, Buccal mucosa.
    stats.table(biom, x = "==Sex", y = "unifrac", by = "==Body Site")

#> Error in FUN(X[[i]], ...): object '==Body Site' not found

    
    # Recall that for distance metrics such as unifrac, the value of
    #  interest (the distance) is calculated from a pair of samples.
    #  Sometimes both samples will be from males, sometimes both from
    #  females, and sometimes one from each.
    # The '==' prefix on '==Sex' and '==Body Site' indicates that only
    #  pairs from the same sex and same body site should be considered.
    # Using the '!=' prefix does the opposite - considering only sample
    #  pairs from different sexes and body sites.
    
    # Removing the '==' prefix generates the complete set of 
    #  comparisons. Note how the test changes from Mann-Whitney to
    #  Kruskal-Wallis. This is because instead of a two group comparison
    #  ('Male', 'Female') we are now comparing three groups 
    #  ('Male', 'Female', 'Female vs Male').
    stats.table(biom, x = "Sex", y = "unifrac", by = "Body Site")

#>                 Body Site   n           Test   p.val  adj.p
#> 1           Buccal mucosa  45 Kruskal-Wallis 0.00319 0.0191
#> 2         Saliva vs Stool  90 Kruskal-Wallis 0.15100 0.3260
#> 3 Buccal mucosa vs Saliva 100 Kruskal-Wallis 0.16300 0.3260
#> 4  Buccal mucosa vs Stool  90 Kruskal-Wallis 0.79700 0.9580
#> 5                  Saliva  45 Kruskal-Wallis 0.80700 0.9580
#> 6                   Stool  36 Kruskal-Wallis 0.95800 0.9580