Subset samples using the BIOM's metadata or taxonomy.

Subset samples using the BIOM's metadata or taxonomy.

# S3 method for BIOM
subset(
  x,
  expr,
  env = parent.frame(),
  drop.na = TRUE,
  refactor = TRUE,
  fast = FALSE
)

Arguments

x
A BIOM object, as returned from read.biom.

expr
Logical expression to run on the metadata or taxonomy (not both) to identify samples or taxa to retain.

env
The environment to search for variables used in expr. (Default: parent.frame())

drop.na
When expr is e.g. Age > 30, should !is.na(Age) be automatically applied too?

refactor
When expr is e.g. `Body Site` %in% c("Stool", "Saliva"), should `Body Site` be redefined as factor(`Body Site`, levels=c('Stool', 'Saliva'))? Applies to non-numeric metadata only.

fast
Should subsetting the phylogenetic tree and sequences be skipped? These slow steps are often not necessary. (Default: FALSE)

Value

A BIOM object.

See also

Examples

    library(rbiom)
    
    ex1 <- subset(hmp50, Age > 30)
    ex2 <- subset(hmp50, Phylum %in% c("Firmicutes", "Actinobacteria"))
    ex3 <- subset(hmp50, `Body Site` %in% c("Saliva", "Stool"))
    ex4 <- subset(hmp50, a == b, list(a = as.name("Body Site"), b ="Saliva"))
    ex5 <- subset(hmp50, Age < 25 & BMI > 22)