Generate a matrix of samples by taxa, at the specified taxonomic rank.

Generate a matrix of samples by taxa, at the specified taxonomic rank.

taxa.rollup(
  biom,
  rank = "OTU",
  map = NULL,
  lineage = FALSE,
  sparse = FALSE,
  long = FALSE,
  md = FALSE,
  safe = FALSE
)

Arguments

biom
A matrix, simple_triplet_matrix, or BIOM object, as returned from read.biom. For matrices, the rows and columns are assumed to be the taxa and samples, respectively.

rank
The taxonomic rank. E.g. “OTU”, “Phylum”, etc. May also be given numerically: 0 for OTU, 1 for the highest level (i.e. Kingdom), and extending to the number of taxonomic ranks encoded in the original biom file. Run taxa.ranks(biom) to fetch the names of all available ranks.

map
A character matrix defining the value that each taxa IDs is assigned for each taxonomic rank. If map=NULL and biom is a BIOM class object, the map will be automatically loaded from biom$taxonomy. map must not be null when biom is a matrix or simple_triplet_matrix. See the example below for an example of map's structure.

lineage
Include all ranks in the name of the taxa. For instance, setting to TRUE will produce Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales. Whereas setting to FALSE (the default) will return simply Coriobacteriales. You want to set this to TRUE if you have genus names (such as Incertae_Sedis) that map to multiple higher level ranks.

sparse
If true, returns a sparse matrix as described by slam::simple_triplet_matrix, otherwise returns a normal R matrix object. Sparse matrices will likely be considerably more memory efficient in this scenario.

long
Pivot the returned data to long format?

FALSE

Each metric has its own column. (Default)

TRUE

"Sample", "Metric" and "Diversity" are the columns returned. Rows are added to attain all combinations of samples x metrics.

md
Include metadata in the output data frame? Options are:

FALSE

Don't include metadata. (Default)

TRUE

Include all metadata.

character vector

Include only the specified metadata columns.

safe
Should autogenerated columns be prefixed with a "." to avoid conflicting with metadata column names? (Default: FALSE) safe=NA can be used with long=FALSE to omit the 'Sample' column entirely, leaving the sample names only as the row names.

Value

A numeric matrix with samples as column names, and taxonomic identifiers as row names. Or a data.frame is md is not NULL.

Examples

    library(rbiom)
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read.biom(infile)
    
    colnames(biom$taxonomy)

#> [1] "Kingdom" "Phylum"  "Class"   "Order"   "Family"  "Genus"  

    
    phyla <- taxa.rollup(biom, 'Phylum')
    phyla[1:4,1:6]

#>       Sample Actinobacteria Bacteroidetes Cyanobacteria Deinococcus Thermus
#> HMP01  HMP01             18           276             0                   0
#> HMP02  HMP02             60           221             0                   0
#> HMP03  HMP03            126           313             0                   0
#> HMP04  HMP04            120           218             0                   0
#>       Firmicutes
#> HMP01       1208
#> HMP02        931
#> HMP03        787
#> HMP04        865

    
    # Custom matrices should be formatted like so:
    counts <- as.matrix(biom$counts)
    map    <- biom$taxonomy
    
    counts[1:3,1:6]

#>          HMP01 HMP02 HMP03 HMP04 HMP05 HMP06
#> UncO2713     0     0     0     0     0     0
#> UncO4101     1     5     6    18     5    15
#> AnmMass2     0     0     0     0     0     0

    map[1:3,1:4]

#>          Kingdom    Phylum           Class            Order            
#> UncO2713 "Bacteria" "Bacteroidetes"  "Bacteroidia"    "Bacteroidales"  
#> UncO4101 "Bacteria" "Firmicutes"     "Clostridia"     "Clostridiales"  
#> AnmMass2 "Bacteria" "Actinobacteria" "Actinobacteria" "Actinomycetales"

    
    phyla <- taxa.rollup(counts, 'Phylum', map=map)
    phyla[1:3,1:6]

#>       Sample Actinobacteria Bacteroidetes Cyanobacteria Deinococcus Thermus
#> HMP01  HMP01             18           276             0                   0
#> HMP02  HMP02             60           221             0                   0
#> HMP03  HMP03            126           313             0                   0
#>       Firmicutes
#> HMP01       1208
#> HMP02        931
#> HMP03        787