Visualize BIOM data with boxplots.

Visualize BIOM data with boxplots.

taxa_boxplot(
  biom,
  x = ".taxa",
  y = tail(taxa.ranks(biom), 1),
  layers = "rls",
  color.by = NULL,
  pattern.by = NULL,
  shape.by = NULL,
  facet.by = NULL,
  xvals = NULL,
  colors = NULL,
  patterns = NULL,
  shapes = NULL,
  facets = NULL,
  taxa = 5,
  p.top = Inf,
  p.adj = "fdr",
  p.label = 0.05,
  perms = 1000,
  ci = 95,
  xlab.angle = "auto",
  other = FALSE,
  safe = FALSE,
  ...
)

Arguments

x

A BIOM object, as returned from read.biom.

layers

"box", "bar" ("r"), "violin", "dot", "strip", "crossbar", "errorbar", "linerange", and "pointrange". Single letter abbreviations are also accepted. For instance, c("box", "dot") is equivalent to c("b", "d") and "bd". Default: "rls".

color.by, pattern.by, shape.by, facet.by

Metadata column to color, pattern, shape, and/or facet by. If that column is a factor, the ordering of levels will be maintained in the plot.

colors, patterns, shapes, facets, xvals

Names of the colors, patterns, shapes, facets, and/or x values to use in the plot. Available values for colors, patterns, and shapes are given by colors(), gridpattern::names_magick, and 0:25, respectively. Use a named character vector to map them to specific factor levels in the metadata. facets and xvals are coerced to unnamed character vectors. If the length of these vectors is less than the values present in their corresponding metadata column, then the data set will be subseted accordingly.

taxa

Which taxa to display. An integer value will show the top n most abundant taxa. A value 0 <= n < 1 will show any taxa with that mean abundance or greater (e.g. 0.1). A character vector of taxon names will show only those taxa. (Default: 10 for heatmaps, 5 otherwise.)

p.top

For RANK ~ . (taxa abundance) plots or BDIV ~ ORD ordination biplots, only display taxa with the most significant differences in abundance. If p.top is >= 1, then the p.top most significant taxa are displayed. If p.top is less than one, all taxa with an adjusted p-value <= p.top are displayed. Recommended to be used in combination with the taxa parameter to set a lower bound on the mean abundance of considered taxa. (Default: Inf)

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. (Default: fdr)

p.label

Minimum adjusted p-value to display on the plot with a bracket. Set to Inf to display all p-values, or -Inf for no brackets. If a numeric vector with more than one value is provided, they will be used as breaks for asterisk notation. (Default: 0.05)

ci

How to calculate min/max of the crossbar, errorbar, linerange, and pointrange layers. Provide a number between 75 and 100 to define a confidence interval's confidence level, commonly 95 or 97.5. Other options are: range, sd (standard deviation), se (standard error), and mad (median absolute deviation). The center mark of crossbar and pointrange represents the mean, except for mad in which case it represents the median. Trendlines require a confidence interval value. Set to NULL to disable. Default: 95

xlab.angle

How to rotate the tick labels on the x-axis. 'auto' (the default), automatically selects a rotation value. 0, 30, and 90 sets the angle to horizontal, angled, and vertical, respectively.

other

Should non-selected taxa be displayed as an "Other" group? (Default: FALSE)

safe

If FALSE, data.frame column names such as ".metric" will be auto-converted to "Metric" to improve human-readability. Conversion if aborted if a conflict is found with a metadata column name. (Default: FALSE)

...

Parameters are matched to formal arguments of ggplot2 functions. Prefixing parameter names with a layer name ensures that a particular parameter is passed to, and only to, that layer. For instance, dot.size = 2 or d.size = 2 ensures only the dotplot layer has its size set to 2.

rank

What rank of taxa to display. E.g. "Phylum", "Genus", etc. Run taxa.ranks() to see all options for a given BIOM object. The default, NULL, selects the lowest level.

Value

A ggplot2 plot. The computed data points and statistics will be attached as attr(p, 'data') and attr(p, 'stats'), respectively. Shapes can also be given as their string values, defined in pch_table here: https://github.com/tidyverse/ggplot2/blob/master/R/geom-point.r . Note that some shapes have a colored outline given by `color`, some are filled with `color`, and some are outlined in `color` and filled with `fill`. See https://blog.albertkuo.me/post/point-shape-options-in-ggplot/ for details. To expand the low end of the y axis, you can set y.trans = "sqrt" or y.trans = "log1p". The former applies a square-root transformation, and the latter plots log(y + 1). Both of these methods work well with data that contains zeroes.

See also

Examples

    library(rbiom)
    
    biom <- rarefy(hmp50)
    taxa_boxplot(biom, x = ".", y = c("Phylum", "Genus"))
#> Error in taxa_boxplot(biom, x = ".", y = c("Phylum", "Genus")): `x` argument to taxa_boxplot must be metadata column name(s) or '.taxa'