Get the taxa abundances.

Get the taxa abundances.

taxa_means(biom, rank = NULL)

Arguments

biom

A BIOM object, as returned from read_biom.

rank

The taxonomic rank to return means for. The default, NULL, return OTU means in the same order as they appear in counts. If not NULL, returned means are ordered from most abundance to least abundant.

Value

A numeric vector, named with the taxa names.

Examples

    library(rbiom)
    
    taxa_means(hmp50, 'Genus') %>% head(10)
#>     Lactobacillus     Streptococcus Corynebacterium 1       Bacteroides 
#>            803.74            491.52            328.66            323.22 
#>    Staphylococcus Propionibacterium        uncultured       Haemophilus 
#>            275.50            251.70            126.54            115.72 
#>       Veillonella         Neisseria 
#>             78.40             50.22