Generate a data.frame of taxa abundance for each sample at the specified taxonomic rank.

Generate a data.frame of taxa abundance for each sample at the specified taxonomic rank.

Usage

taxa_table(
  biom,
  rank = "OTU",
  taxa = NULL,
  map = NULL,
  lineage = FALSE,
  md = FALSE,
  unc = "singly",
  other = FALSE,
  safe = FALSE
)

Arguments

biom

A matrix, simple_triplet_matrix, or BIOM object, as returned from read_biom. For matrices, the rows and columns are assumed to be the taxa and samples, respectively.

rank

The taxonomic rank. E.g. rank = "OTU", rank = "Phylum", etc. May also be given numerically: 0 for OTU, 1 for the highest level (i.e. Kingdom), and extending to the number of taxonomic ranks encoded in the original biom file. Run taxa_ranks(biom) to fetch the names of all available ranks.

taxa

Limit the number of taxa returned in the matrix. Depending on the value type, a different filter is applied.

NULL

Retain all taxa. (The default.)

Integer >= 1

The N most abundant taxa are retained. Example: taxa_matrix(biom, 'Genus', 5) returned the five most abundant genera.

Numeric between 0 and 1

Taxa are retained if their abundance is greater than or equal to N. Example: taxa_matrix(biom, 'Phylum', 0.1) returns all phyla with a relative abundance of at least 10 %.

Character vector

Only these taxa names are retained. Example: taxa_matrix(biom, 'Phylum', c("Firmicutes", "Bacteroidetes")) will only retain those two phyla.

map

A character matrix defining the value that each taxa IDs is assigned for each taxonomic rank. If map=NULL and biom is a BIOM class object, the map will be automatically loaded from biom$taxonomy. map must not be null when biom is a matrix or simple_triplet_matrix. See the example below for an example of map's structure.

lineage

Include all ranks in the name of the taxa. For instance, setting to TRUE will produce Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales. Whereas setting to FALSE (the default) will return simply Coriobacteriales. You want to set this to TRUE if you have genus names (such as Incertae_Sedis) that map to multiple higher level ranks.

md

Include metadata in the output data frame? Options are:

FALSE

Don't include metadata. (Default)

TRUE

Include all metadata.

character vector

Include only the specified metadata columns.

unc

How to handle unclassified, uncultured, and similarly ambiguous taxa names. The default, "singly"/"s", replaces them with the OTU name. Other options are "grouped"/"g", which replaces them with a higher rank's name, "drop"/"d" which excludes them from the result, and "asis"/"a" to not check/modify any taxa names.

other

Sum all non-itemized taxa into an "Other" taxa. The default, FALSE, only returns taxa matched by the `taxa` argument. Specifying TRUE adds "Other" to the returned set. A string can also be given to imply TRUE, but with that value as the name to use in place of "Other".

safe

Should autogenerated columns be prefixed with a "." to avoid conflicting with metadata column names? (Default: FALSE) safe=NA can be used with long=FALSE to omit the 'Sample' column entirely, leaving the sample names only as the row names.

Value

A numeric matrix with samples as column names, and taxonomic identifiers as row names. Or a data.frame if metadata is requested with md.

Examples

    library(rbiom)
    
    taxa_ranks(hmp50)
#> [1] "Kingdom" "Phylum"  "Class"   "Order"   "Family"  "Genus"  
    
    phyla <- taxa_table(hmp50, 'Phylum')
    phyla[1:4,1:6]
#> Error in `[.data.frame`(phyla, 1:4, 1:6): undefined columns selected
    
    # Custom matrices should be formatted like so:
    counts <- counts(hmp50)
    map    <- taxonomy(hmp50)
    
    counts[1:3,1:6]
#>          HMP01 HMP02 HMP03 HMP04 HMP05 HMP06
#> UncO2713     0     0     0     0     0     0
#> UncO4101     1     5     6    18     5    15
#> AnmMass2     0     0     0     0     0     0
    map[1:3,1:4]
#>          Kingdom    Phylum           Class            Order            
#> UncO2713 "Bacteria" "Bacteroidetes"  "Bacteroidia"    "Bacteroidales"  
#> UncO4101 "Bacteria" "Firmicutes"     "Clostridia"     "Clostridiales"  
#> AnmMass2 "Bacteria" "Actinobacteria" "Actinobacteria" "Actinomycetales"
    
    phyla <- taxa_table(counts, 'Phylum', map=map)
    phyla[1:3,1:6]
#> Error in `[.data.frame`(phyla, 1:3, 1:6): undefined columns selected