Provides a 'treedata' S4 object for use in ggtree functions.

Provides a 'treedata' S4 object for use in ggtree functions.

tree_data(biom, reads = TRUE, clades = TRUE)

Arguments

biom

A BIOM object, as returned from read_biom.

reads

Include a 'reads' column indicating the sum of taxa observations belonging to each node/leaf. Default: TRUE.

clades

Notate top-most monophyletic nodes. The default, TRUE, adds a column for every rank in the biom object. A character vector of ranks can also be passed in. If the vector is named, then those names are used for naming the columns in the returned treedata object. Set to NULL to not return any clade notations.

Examples

    library(rbiom)
    
    tree_data(hmp50)
#> 'treedata' S4 object'.
#> 
#> ...@ phylo:
#> 
#> Phylogenetic tree with 490 tips and 489 internal nodes.
#> 
#> Tip labels:
#>   EschC738, Unc92490, Unc02pjp, Unc26984, Unc41900, AntBa370, ...
#> 
#> Rooted; includes branch lengths.
#> 
#> with the following features available:
#>   'reads', 'Kingdom', 'Phylum', 'Class', 'Order', 'Family', 'Genus', 'OTU'.
#> 
#> # The associated data tibble abstraction: 979 × 11
#> # The 'node', 'label' and 'isTip' are from the phylo tree.
#>     node label    isTip reads Kingdom Phylum Class Order      Family Genus OTU  
#>    <dbl> <chr>    <lgl> <dbl> <chr>   <chr>  <chr> <chr>      <chr>  <chr> <chr>
#>  1     1 EschC738 TRUE    100 NA      NA     NA    Enterobac… Enter… Esch… Esch…
#>  2     2 Unc92490 TRUE    746 NA      NA     NA    NA         NA     Haem… Unc9…
#>  3     3 Unc02pjp TRUE    242 NA      NA     NA    NA         NA     Acti… Unc0…
#>  4     4 Unc26984 TRUE   5040 NA      NA     NA    NA         NA     Haem… Unc2…
#>  5     5 Unc41900 TRUE      1 NA      NA     NA    Enterobac… Enter… Prov… Unc4…
#>  6     6 AntBa370 TRUE     43 NA      NA     NA    NA         NA     NA    AntB…
#>  7     7 AntSp625 TRUE      1 NA      NA     NA    NA         NA     NA    AntS…
#>  8     8 MxlSpe11 TRUE      8 NA      NA     NA    NA         NA     NA    MxlS…
#>  9     9 FuoMancu TRUE      1 NA      NA     NA    NA         NA     NA    FuoM…
#> 10    10 PeuS1830 TRUE     18 NA      NA     NA    Pseudomon… Pseud… Pseu… PeuS…
#> # … with 969 more rows