Compute Weighted and Unweighted UniFrac distance matrices.

This is the function called internally by beta.div, but is made visible here so you can use it with matrices and trees without having to first convert them to BIOM objects.

unifrac(biom, weighted = TRUE, tree = NULL, progressbar = FALSE)

Arguments

biom
A matrix, simple_triplet_matrix, or BIOM object, as returned from read.biom. For matrices, the rows and columns are assumed to be the taxa and samples, respectively.

weighted
Use weighted UniFrac, which takes abundance into account rather than simply presence/absence.

tree
A phylo object providing a phylogenetic tree for the taxa names in biom. If tree=NULL, then the tree will be loaded from biom, if encoded there.

progressbar
Whether to display a progress bar and status messages (logical). Will automatically tie in with shiny if run within a shiny session.

Value

A distance matrix of class dist.

Examples

    library(rbiom)

    infile <- system.file("extdata", "hmp50.biom", package = "rbiom")
    biom <- read.biom(infile)
    biom <- select(biom, 1:10)

    dm <- unifrac(biom)

#> Error in par_unifrac(counts, tree, ifelse(weighted, 1L, 0L)): object '_rbiom_par_unifrac' not found

    plot(hclust(dm), cex=.8)

#> Error in hclust(dm): object 'dm' not found

    as.matrix(dm)[1:4,1:4]

#> Error in as.matrix(dm): object 'dm' not found


    # Using a custom matrix and tree
    mtx <- matrix(sample.int(12*20), ncol=20)
    dimnames(mtx) <- list(LETTERS[1:12], letters[1:20])
    tree <- ape::as.phylo(hclust(dist(mtx)))

    dm <- unifrac(mtx, tree=tree)

#> Error in par_unifrac(counts, tree, ifelse(weighted, 1L, 0L)): object '_rbiom_par_unifrac' not found

    as.matrix(dm)[1:4,1:4]

#> Error in as.matrix(dm): object 'dm' not found