Write data and summary information to a Microsoft Excel-compatible workbook.

Write data and summary information to a Microsoft Excel-compatible workbook.

Usage

write_xlsx(biom, file = NULL, depth = NULL, seed = 0, unc = "asis")

Arguments

biom

The rbiom object to save to the file.

file

Path to the output xlsx file.

depth

Depth to rarefy to. See sample_rarefy() function for details. depth = NULL auto-selects a rarefaction level. depth = 0 disables rarefaction. Only use depth with 'OTU table' data and integer count values.

seed

Random seed to use in rarefying. See sample_rarefy() function for details.

unc

How to handle unclassified, uncultured, and similarly ambiguous taxa names. Options are:

"singly" - Replaces them with the OTU name.

"grouped" - Replaces them with a higher rank's name.

"drop" - Excludes them from the result.

"asis" - To not check/modify any taxa names.

Default: "singly"

Abbreviations are allowed.

Value

The normalized filepath that was written to (invisibly).

Note

Any data frame, matrix, or dist attributes on biom will be included as separate worksheets. An attribute named 'Reads Per Step' is treated specially and merged with the usual 'Reads Per Sample' tab.

See also

Examples

    library(rbiom) 
    
    if (FALSE) {
      
      biom <- sample_select(hmp50, 1:10) %>% sample_rarefy()
      
      attr(biom, "Weighted UniFrac")   <- bdiv_distmat(biom, 'unifrac')
      attr(biom, "Unweighted Jaccard") <- bdiv_distmat(biom, 'jaccard', weighted=FALSE)
      
      outfile <- write_xlsx(biom, tempfile(fileext = ".xlsx"))
    }