Test alpha diversity differences for significance.

Test alpha diversity differences for significance.

Usage

adiv_stats(
  biom,
  stat.by = NULL,
  regr = NULL,
  adiv = "Shannon",
  test = ifelse(is.null(regr), "means", "trends"),
  model = "lm",
  split.by = NULL,
  level = 0.95,
  p.adj = "fdr"
)

Arguments

biom

An rbiom-class object, or data coercible with as_rbiom().

stat.by

The categorical metadata field defining the statistical groups. Required.

regr

To run a regression analysis, set regr to the numeric metadata field with the "x-axis" values. Leaving regr=NULL will generate boxplot-like statistics; when non-NULL, corrplot-like statistics will be returned. Default: NULL

adiv

Alpha diversity metric(s) to use. Options are: "OTUs", "Shannon", "Chao1", "Simpson", and/or "InvSimpson". The shortcut adiv="all" also exists. Default: "Shannon"

Multiple values allowed. Non-ambiguous abbreviations are allowed.

test

The statistic of interest. An overview of options can be found in the "Statistical Tests" section below. One of "predict", "terms", "fit", "means", "trends", "es_means", "es_trends", "pw_means", or "pw_trends". Default: ifelse(is.null(regr), "means", "trends")

model

What type of trend model to fit to the data. Options are: "lm" (linear), "log" (logarithmic), or "gam" (generalized additive). See the "Model Options" section below for additional details. Default: "lm"

split.by

Name(s) of metadata columns that the data should be split by prior to any calculations. Default: NULL

level

The confidence level for calculating a confidence interval. Default: 0.95

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: "fdr".

Value

A tibble data frame with summary statistics.

The R code or generating these statistics is in $code, and the object history is in $history.

Statistical Tests

When regr is NULL:

"means" - Considers all groups at once using the Kruskal-Wallis non-parametric test: stats::kruskal.test().

"pw_means" - Pairwise comparison of stat.by groups using the Wilcox rank sum (aka Mann-Whitney) non-parametric test: stats::wilcox.test().

When regr is not NULL:

"predict" - Augments original data with fitted information. See broom::augment.lm().

"terms" - Summary information about the model's terms, including p-value, r-squared, AIC, BIC, and confidence interval at level. See broom::tidy.lm().

"fit" - Goodness of fit measures, p-values, and more for the overall model. See broom::glance.lm().

"means" - The estimated marginal mean (EMM) for each stat.by group, along with confidence intervals (CI), standard errors (SE), t-ratios, and p-values testing for mean = zero. See emmeans::emmeans() and emmeans::summary.emmGrid().

"trends" - The trendline slope for each stat.by group, along with CI and SE and p-value testing for slope = zero. See emmeans::emtrends() and emmeans::summary.emmGrid().

"pw_means" - Pairwise means. All stat.by groups are compared to each other and the difference in means is estimated along with SE, t-ratios, and p-values testing if the two means are the same. See emmeans::pairs.emmGrid().

"pw_trends" - Pairwise trends. As above, but comparing trendline slopes instead of means.

"es_means" - Effect sizes for pairwise differences of means, including SE and CI. See emmeans::eff_size().

"es_trends" - Effect sizes for pairwise differences of slopes, including SE and CI. See emmeans::eff_size().

Model Options

The predefined options are:

"lm" - Linear model: stats::lm(formula = y ~ x).)

"log" - Logarithmic model: stats::lm(formula = y ~ log(x)).

"gam" - Generalized additive model: mgcv::gam(formula = y ~ s(x, bs = "cs"), method = "REML").

You can alternatively provide a list of length two where the first element is a character vector of length 1 naming a function, and the second element is a list of arguments to pass to that function. One of the function's arguments must be named 'formula'. For example, model = list("stats::lm", list(formula = y ~ x)).

See also

Examples

    library(rbiom) 
    
    biom <- sample_rarefy(hmp50)
      
    adiv_stats(biom, stat.by = "Body Site")
#> # Test:     kruskal.test(.Shannon ~ `Body Site`).
#> # A tibble: 1 × 5
#>      .n .stat   .df       .p.val       .adj.p
#>   <int> <dbl> <int>        <dbl>        <dbl>
#> 1    49  38.7     4 0.0000000804 0.0000000804
      
    adiv_stats(biom, stat.by = "Sex", test = "pw_means")
#> # Test:     wilcox.test(.Shannon ~ Sex).
#> # A tibble: 1 × 8
#>   Sex              .n .stat .estimate .lower .upper .p.val .adj.p
#>   <fct>         <dbl> <dbl>     <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1 Female - Male    49   167    -0.781  -1.44 -0.183 0.0149 0.0149
      
    adiv_stats(biom, stat.by = "Body Site", regr = "Age")
#> # Test:     Estimated marginal means of linear trends.
#> # Model:    stats::lm(.Shannon ~ Age * `Body Site`)
#> # A tibble: 5 × 9
#>   `Body Site`    .trend    .se   .df   .lower .upper .t.ratio .p.val .adj.p
#>   <fct>           <dbl>  <dbl> <dbl>    <dbl>  <dbl>    <dbl>  <dbl>  <dbl>
#> 1 Mid vagina     0.0450 0.0265    39 -0.00858 0.0985    1.70  0.0973  0.487
#> 2 Buccal mucosa  0.0661 0.0554    39 -0.0460  0.178     1.19  0.240   0.601
#> 3 Saliva         0.0228 0.0356    39 -0.0492  0.0948    0.640 0.526   0.737
#> 4 Anterior nares 0.0108 0.0266    39 -0.0430  0.0646    0.406 0.687   0.737
#> 5 Stool          0.0147 0.0434    39 -0.0731  0.103     0.339 0.737   0.737
    
    adiv_stats(biom, stat.by = "Body Site", split.by = "Sex")
#> # Test:     kruskal.test(.Shannon ~ `Body Site`).
#> # A tibble: 2 × 6
#>   Sex       .n .stat   .df    .p.val   .adj.p
#>   <fct>  <int> <dbl> <int>     <dbl>    <dbl>
#> 1 Female    30  24.1     4 0.0000779 0.000156
#> 2 Male      19  13.8     3 0.00325   0.00325