Calculate the alpha diversity of each sample.

Calculate the alpha diversity of each sample.

Usage

adiv_table(biom, rarefy = FALSE, adiv = "Shannon", md = ".all")

Arguments

biom

An rbiom-class object, or data coercible with as_rbiom().

rarefy

Control how/whether rarefactions are done prior to alpha diversity computations. Options are:

FALSE - Use each sample's current set of observations without applying any rarefaction.

TRUE - Automatically select and apply a single rarefaction.

"multi", "multi_log", "multi_even" - Automatically select and apply multiple rarefactions. "multi" provides "multi_log" at the low end and "multi_even" at the high end.

integer vector - Rarefy at the specified depth(s).

Default: FALSE

adiv

Alpha diversity metric(s) to use. Options are: "OTUs", "Shannon", "Chao1", "Simpson", and/or "InvSimpson". Set adiv=".all" to use all metrics. Default: "Shannon"

Multiple values allowed. Non-ambiguous abbreviations are allowed.

md

A character vector naming the metadata fields to include in the output data frame, or '.all' to include all metadata fields. Default: '.all'

Value

A data frame of alpha diversity values.

Each combination of sample/depth/adiv has its own row.

Column names are .sample, .depth, .adiv, and .diversity, followed by any metadata fields requested by md.

See also

Examples

    library(rbiom) 
    
    biom <- sample_select(hmp50, 1:6)
    
    adiv_table(biom)
#> # A tibble: 6 × 8
#>   .sample .depth .adiv   .diversity   Age   BMI `Body Site`   Sex   
#>   <chr>    <dbl> <fct>        <dbl> <dbl> <dbl> <fct>         <fct> 
#> 1 HMP01     1660 Shannon       1.74    22    20 Buccal mucosa Female
#> 2 HMP02     1371 Shannon       2.59    24    23 Buccal mucosa Male  
#> 3 HMP03     1353 Shannon       2.95    28    26 Saliva        Male  
#> 4 HMP04     1895 Shannon       3.26    25    23 Saliva        Male  
#> 5 HMP05     3939 Shannon       1.46    27    24 Buccal mucosa Female
#> 6 HMP06     4150 Shannon       3.09    32    25 Saliva        Male  
    
    adiv_table(sample_rarefy(biom), adiv = ".all", md = NULL)
#> # A tibble: 30 × 4
#>    .sample .depth .adiv   .diversity
#>    <fct>    <dbl> <fct>        <dbl>
#>  1 HMP01     1353 OTUs         46   
#>  2 HMP02     1353 OTUs         75   
#>  3 HMP03     1353 OTUs         75   
#>  4 HMP04     1353 OTUs         77   
#>  5 HMP05     1353 OTUs         40   
#>  6 HMP06     1353 OTUs         86   
#>  7 HMP01     1353 Shannon       1.74
#>  8 HMP02     1353 Shannon       2.59
#>  9 HMP03     1353 Shannon       2.95
#> 10 HMP04     1353 Shannon       3.25
#> # ℹ 20 more rows