Reduce a distance matrix to two or three dimensions

Reduce a distance matrix to two or three dimensions

Usage

bdiv_ord_table(
  biom,
  dist = "Bray-Curtis",
  ord = "UMAP",
  weighted = TRUE,
  tree = NULL,
  md = NULL,
  k = 2,
  split.by = NULL,
  stat.by = NULL,
  seed = 0,
  perms = 1000,
  rank = NULL,
  taxa = 5,
  p.adj = "fdr",
  p.top = 5,
  ...
)

Arguments

biom

A BIOM object, as returned from read_biom. Alternatively, a distance matrix, such as from bdiv_distmat.

dist

The distance algorithm to use. Options are: "Bray-Curtis", "Manhattan", "Euclidean", "Jaccard", and "UniFrac". A phylogenetic tree must be present in biom or explicitly provided via tree= to use the UniFrac methods. Ignored if biom is a distance matrix. Multiple values allowed. Default: "Bray-Curtis".

ord

Method for reducing the dimensionality of a distance matrix. Multiple values allowed. Options are:

"UMAP" (default)

Uniform manifold approximation and projection, via umap.

"PCoA"

Principal coordinate analysis, via pcoa.

"NMDS"

Nonmetric multidimensional scaling, via metaMDS.

"tSNE"

t-distributed stochastic neighbor embedding, via tsne.

weighted

Take relative abundances into account. When weighted=FALSE, only presence/absence is considered. Ignored if biom is a distance matrix. Multiple values allowed. (Default: TRUE)

tree

A phylo object representing the phylogenetic relationships of the taxa in biom. Will be taken from the tree embedded in the biom object if not explicitly specified. Only required for computing UniFrac distance matrices. Ignored if biom is a distance matrix.

md

Include metadata in the output data frame? Ignored if biom is a distance matrix. Options are:

NULL

Don't include metadata. (Default)

TRUE

Include all metadata.

character vector

Include only the specified metadata columns.

k

Number of dimensions to return. (Default: 2)

split.by

Name(s) of metadata columns that the data should be split by prior to calculating distance matrices, ordinations, and statistics.

stat.by, seed, perms

Passthrough parameters to bdiv_distmat() for computing adonis statistics. (Default: stat.by=NULL, seed=0, perms=1000)

rank

[Biplot] What rank of taxa to display (e.g. "Phylum"), or NULL for no biplot. Run taxa_ranks() to see all options for a given BIOM object. (Default: NULL)

taxa

[Biplot] Which taxa to display. An integer value will return the top n most abundant taxa. A value 0 <= n < 1 will return any taxa with that mean abundance or greater (e.g. 0.1). A character vector of taxon names will show only those taxa. (Default: 5)

p.adj

[Biplot] Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. (Default: fdr)

p.top

[Biplot] Only returns taxa with the most significant differences in abundance. If p.top is >= 1, then the p.top most significant taxa are displayed. If p.top is less than one, all taxa with an adjusted p-value <= p.top are displayed. Recommended to be used in combination with the taxa parameter to set a lower bound on the mean abundance of considered taxa. (Default: Inf)

...

Additional arguments to pass on to pcoa, metaMDS, or tsne.

Value

A data.frame with columns .axis.1, .axis.2, ..., axis.k as well as columns given by md, split.by, and stat.by. Sample IDs are in column '.sample' and the metric used in column '.metric'. For pcoa ordinations,

attr(, 'eig') will contain the eigenvalues useful for construction of "..

attr(, 'stats_raw') and attr(,'stats_tbl') contain the raw and tabular adonis statistics when stat.by is set. Distance matrices are in attr(, 'dm'). When rank is non-NULL, or taxa is a list of taxon names, a data.frame of biplot coordinates will be given by attr(,'biplot').

Examples

    library(rbiom)
    
    ord <- bdiv_ord_table(hmp50, "bray-curtis", "pcoa")
    head(ord)
#>      .axis.1      .axis.2 .sample  .weight       .dist .ord
#> 1 -0.3575401  0.002362419   HMP01 Weighted Bray-Curtis PCoA
#> 2 -0.4203019  0.010404374   HMP02 Weighted Bray-Curtis PCoA
#> 3 -0.3819160  0.013483401   HMP03 Weighted Bray-Curtis PCoA
#> 4 -0.3876548  0.004854090   HMP04 Weighted Bray-Curtis PCoA
#> 5 -0.4154969 -0.011498409   HMP05 Weighted Bray-Curtis PCoA
#> 6 -0.4135847 -0.007029588   HMP06 Weighted Bray-Curtis PCoA