Test beta diversity for associations with metadata.

A convenience wrapper for bdiv_table() + stats_table().

Usage

bdiv_stats(
  biom,
  regr = NULL,
  stat.by = NULL,
  bdiv = "Bray-Curtis",
  weighted = TRUE,
  tree = NULL,
  within = NULL,
  between = NULL,
  split.by = NULL,
  trans = "none",
  test = "emmeans",
  fit = "gam",
  at = NULL,
  level = 0.95,
  alt = "!=",
  mu = 0,
  p.adj = "fdr"
)

Arguments

biom

An rbiom object, such as from as_rbiom(). Any value accepted by as_rbiom() can also be given here.

regr

Dataset field with the x-axis (independent; predictive) values. Must be numeric. Default: NULL

stat.by

Dataset field with the statistical groups. Must be categorical. Default: NULL

bdiv

Beta diversity distance algorithm(s) to use. Options are: "Bray-Curtis", "Manhattan", "Euclidean", "Jaccard", and "UniFrac". For "UniFrac", a phylogenetic tree must be present in biom or explicitly provided via tree=. Default: "Bray-Curtis"

Multiple/abbreviated values allowed.

weighted

Take relative abundances into account. When weighted=FALSE, only presence/absence is considered. Default: TRUE

Multiple values allowed.

tree

A phylo object representing the phylogenetic relationships of the taxa in biom. Only required when computing UniFrac distances. Default: biom$tree

within, between

Dataset field(s) for intra- or inter- sample comparisons. Alternatively, dataset field names given elsewhere can be prefixed with '==' or '!=' to assign them to within or between, respectively. Default: NULL

split.by

Dataset field(s) that the data should be split by prior to any calculations. Must be categorical. Default: NULL

trans

Transformation to apply. Options are: c("none", "rank", "log", "log1p", "sqrt"). "rank" is useful for correcting for non-normally distributions before applying regression statistics. Default: "none"

test

Method for computing p-values: 'wilcox', 'kruskal', 'emmeans', or 'emtrends'. Default: 'emmeans'

fit

How to fit the trendline. 'lm', 'log', or 'gam'. Default: 'gam'

at

Position(s) along the x-axis where the means or slopes should be evaluated. Default: NULL, which samples 100 evenly spaced positions and selects the position where the p-value is most significant.

level

The confidence level for calculating a confidence interval. Default: 0.95

alt

Alternative hypothesis direction. Options are '!=' (two-sided; not equal to mu), '<' (less than mu), or '>' (greater than mu). Default: '!='

mu

Reference value to test against. Default: 0

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: "fdr"

Value

A tibble data.frame with fields from the table below. This tibble object provides the $code operator to print the R code used to generate the statistics.

Field	Description

.mean	Estimated marginal mean. See emmeans::emmeans().

.mean.diff	Difference in means.

.slope	Trendline slope. See emmeans::emtrends().

.slope.diff	Difference in slopes.

.h1	Alternate hypothesis.

.p.val	Probability that null hypothesis is correct.

.adj.p	.p.val after adjusting for multiple comparisons.

.effect.size	Effect size. See emmeans::eff_size().

.lower	Confidence interval lower bound.

.upper	Confidence interval upper bound.

.se	Standard error.

.n	Number of samples.

.df	Degrees of freedom.

.stat	Wilcoxon or Kruskal-Wallis rank sum statistic.

.t.ratio	.mean / .se

.r.sqr	Percent of variation explained by the model.

.adj.r	.r.sqr, taking degrees of freedom into account.

.aic	Akaike Information Criterion (predictive models).

.bic	Bayesian Information Criterion (descriptive models).

.loglik	Log-likelihood goodness-of-fit score.

.fit.p	P-value for observing this fit by chance.

See also

Examples

    library(rbiom)
    
    biom <- rarefy(hmp50)
      
    bdiv_stats(biom, stat.by = "Sex", bdiv = c("bray", "unifrac"))[,1:7]
#> # Model:    gam(.distance ~ Sex, method = "REML")
#> # A tibble: 6 × 7
#>   .bdiv       Sex                   .mean.diff .h1    .p.val .adj.p .effect.size
#>   <fct>       <chr>                      <dbl> <fct>   <dbl>  <dbl>        <dbl>
#> 1 Bray-Curtis Male - Female           -0.0559  != 0  0.00711 0.0353      -0.244 
#> 2 Bray-Curtis Male - Male vs Female   -0.0459  != 0  0.0118  0.0353      -0.220 
#> 3 UniFrac     Female - Male vs Fem…   -0.0238  != 0  0.0901  0.180       -0.108 
#> 4 UniFrac     Male - Male vs Female   -0.0199  != 0  0.263   0.395       -0.0976
#> 5 Bray-Curtis Female - Male vs Fem…    0.0101  != 0  0.469   0.563        0.0461
#> 6 UniFrac     Male - Female            0.00392 != 0  0.850   0.850        0.0170
    
    bdiv_stats(biom, stat.by = "Body Site", split.by = "==Sex")[,1:6]
#> Error in validate_var_choices("xi", choices, evar = evar): Invalid `split.by` option(s): ==Sex
#> Choices are: .sample1, .sample2, .weighted, .bdiv, .distance, Body Site, Sex