Make a data.frame of distances between samples.

Make a data.frame of distances between samples.

bdiv_table(
  biom,
  method = "bray-curtis",
  weighted = TRUE,
  tree = NULL,
  md = FALSE,
  safe = FALSE,
  stat.by = NULL,
  seed = 0,
  perms = 999
)

Arguments

biom

A matrix, simple_triplet_matrix, or BIOM object, as returned from read_biom. For matrices, the rows and columns are assumed to be the taxa and samples, respectively.

method

The distance algorithm to use. Options are: "bray-curtis", "manhattan", "euclidean", "jaccard", and "unifrac". Non-ambiguous abbreviations of the method names are also accepted. A phylogenetic tree must be present in biom or explicitly provided via tree= to use the UniFrac methods. (Default: "bray-curtis")

weighted

Take relative abundances into account. When weighted=FALSE, only presence/absence is considered. (Default: TRUE)

tree

A phylo object representing the phylogenetic relationships of the taxa in biom. Will be taken from the tree embedded in the biom object if not explicitly specified. Only required for computing UniFrac distance matrices. (Default: NULL)

md

Include metadata in the output data frame? Options are:

FALSE

Don't include metadata. (Default)

TRUE

Include all metadata.

character vector

Include only the specified metadata columns. Column names can be prefixed with == or != to indicate that only within or between groupings, respectively, are to be kept. See examples below.

safe

Should autogenerated columns be prefixed with a "." to avoid conflicting with metadata column names? (Default: FALSE)

stat.by

Specify a categorical metadata column name to compute adonis statistics and return results as attributes named 'stats_raw' and 'stats_tbl'. (Default: NULL)

seed

Random seed for adonis permutations. (Default: 0)

perms

Number of random permutations to use in adonis calcuation. (Default: 999)

Value

A data.frame.

Examples

    library(rbiom)
    
    # Return in long format with metadata
    biom <- select(hmp50, 18:21)
    bdiv_table(biom, 'unifrac', md = c("Body Site", "Sex"))
#>    Sample1 Sample2  Distance       Body Site            Sex
#> 5    HMP18   HMP19 0.6651627 Saliva vs Stool Female vs Male
#> 9    HMP18   HMP20 0.6810017 Saliva vs Stool Female vs Male
#> 10   HMP19   HMP20 0.4183059           Stool         Female
#> 13   HMP18   HMP21 0.7170374 Saliva vs Stool           Male
#> 14   HMP19   HMP21 0.3896741           Stool Female vs Male
#> 15   HMP20   HMP21 0.1490926           Stool Female vs Male
    
    # Only look at distances among the stool sample
    bdiv_table(biom, 'unifrac', md = c("==Body Site", "Sex"))
#>    Sample1 Sample2  Distance Body Site            Sex
#> 10   HMP19   HMP20 0.4183059     Stool         Female
#> 14   HMP19   HMP21 0.3896741     Stool Female vs Male
#> 15   HMP20   HMP21 0.1490926     Stool Female vs Male
    
    # Or between males and females
    bdiv_table(biom, 'unifrac', md = c("Body Site", "!=Sex"))
#>    Sample1 Sample2  Distance       Body Site            Sex
#> 5    HMP18   HMP19 0.6651627 Saliva vs Stool Female vs Male
#> 9    HMP18   HMP20 0.6810017 Saliva vs Stool Female vs Male
#> 14   HMP19   HMP21 0.3896741           Stool Female vs Male
#> 15   HMP20   HMP21 0.1490926           Stool Female vs Male