Convert to a data.frame with metadata and diversity/abundance values.

Convert to a data.frame with metadata and diversity/abundance values.

distill(biom, metric, weighted = TRUE, long = TRUE, md = TRUE)

Arguments

biom
A BIOM object, as returned from read.biom.

metric
The diversity/abundance values of interest. Options are:

Alpha Diversity Metrics (one or more)

OTUs, Shannon, Chao1, Simpson, and/or InvSimpson. You may also set metric = "Diversity" to get all five metrics.

Beta Diversity Metrics (one only)

manhattan, euclidean, bray-curtis, jaccard, or unifrac. Use in combination with the weighted parameter. Metadata column names can be prefixed with == or != to limit distance calculations to within or between, respectively, those categories. See examples below. Setting metric = "Distance" will use unifrac if a phylogenetic tree is present, or bray-curtis otherwise.

Taxa Abundances (one only)

Kingdom, Phylum, Class, Order, Family, Genus, Species, Strain, or OTU. Supported ranks will vary by biom. Run taxa.ranks(biom) to see the available options. Specifying metric = "Abundance" will default to the most precise rank possible.

weighted
When metric is a beta diversity metric, should it be run in in weighted mode? (Default: TRUE)

long
Pivot the returned data to long format?

FALSE

Each metric has its own column. (Default)

TRUE

"Sample", "Metric" and "Diversity" are the columns returned. Rows are added to attain all combinations of samples x metrics.

md
Include metadata in the output data frame? Options are:

FALSE

Don't include metadata. (Default)

TRUE

Include all metadata.

character vector

Include only the specified metadata columns.

Value

A data.frame object. The first column will be named Sample, and when possible, the rownames will hold the sample name as well. For beta diversity metrics the first two columns will be named Sample1 and Sample2. The remaining columns will be largely dependent on metric, long, and md. See examples below.

See also

Examples

    library(rbiom) 
    
    infile <- system.file("extdata", "hmp50.bz2", package = "rbiom")
    biom <- read.biom(infile)
    
    distill(biom, "UniFrac", md=c("==Body Site", "!=Sex")) %>% head()

#>     Sample1 Sample2  Distance     Body Site            Sex
#> 51    HMP01   HMP02 0.1950253 Buccal mucosa Female vs Male
#> 202   HMP02   HMP05 0.2509451 Buccal mucosa Female vs Male
#> 301   HMP01   HMP07 0.1395698 Buccal mucosa Female vs Male
#> 305   HMP05   HMP07 0.1057077 Buccal mucosa Female vs Male
#> 353   HMP03   HMP08 0.3257210        Saliva Female vs Male
#> 354   HMP04   HMP08 0.2328171        Saliva Female vs Male

    
    distill(biom, c("Shannon", "OTUs"), md=c("BMI", "Sex")) %>% head()

#>   Sample  Metric Diversity BMI    Sex
#> 1  HMP01 Shannon  1.741153  20 Female
#> 2  HMP01    OTUs 49.000000  20 Female
#> 3  HMP02 Shannon  2.587403  23   Male
#> 4  HMP02    OTUs 75.000000  23   Male
#> 5  HMP03 Shannon  2.950982  26   Male
#> 6  HMP03    OTUs 75.000000  26   Male

    
    distill(biom, "Phylum", long=FALSE, md=FALSE)[1:4,1:4]

#>       Sample __Actinobacteria __Bacteroidetes __Cyanobacteria
#> HMP01  HMP01               18             276               0
#> HMP02  HMP02               60             221               0
#> HMP03  HMP03              126             313               0
#> HMP04  HMP04              120             218               0

    
    distill(biom, "Phylum", long=FALSE, md=c("Age", "Body Site"))[1:4,1:6]

#>   Sample Age     Body Site __Actinobacteria __Bacteroidetes __Cyanobacteria
#> 1  HMP01  22 Buccal mucosa               18             276               0
#> 2  HMP02  24 Buccal mucosa               60             221               0
#> 3  HMP03  28        Saliva              126             313               0
#> 4  HMP04  25        Saliva              120             218               0