Convert to a data.frame with metadata and diversity/abundance values.

Convert to a data.frame with metadata and diversity/abundance values.

distill(
  biom,
  metric,
  weighted = TRUE,
  rarefy = FALSE,
  long = TRUE,
  md = TRUE,
  safe = FALSE
)

Arguments

biom

A BIOM object, as returned from read_biom.

metric

The diversity/abundance values of interest. Options are:

Alpha Diversity Metrics

OTUs, Shannon, Chao1, Simpson, and/or InvSimpson.

Beta Diversity Metrics

manhattan, euclidean, bray-curtis, jaccard, or unifrac. Use in combination with the weighted parameter. Metadata column names can be prefixed with == or != to limit distance calculations to within or between, respectively, those categories. See examples below.

Taxa Abundances

Kingdom, Phylum, Class, Order, Family, Genus, Species, Strain, or OTU. Supported ranks will vary by biom. Run taxa_ranks(biom) to see the available options.

weighted

When metric is a beta diversity metric, should it be run in in weighted mode? (Default: TRUE)

rarefy

Should the dataset be rarefied first? When metric is an alpha diversity metric, this 'rarefy' parameter is passed on directly to adiv_table(). (Default: FALSE)

long

Pivot the returned data to long format?

FALSE

Each metric has its own column.

TRUE

"Sample", "Metric" and "Diversity" are the columns returned. Rows are added to attain all combinations of samples x metrics. (Default)

md

Include metadata in the output data frame? Options are:

FALSE

Don't include metadata. (Default)

TRUE

Include all metadata.

character vector

Include only the specified metadata columns.

safe

Should autogenerated columns be prefixed with a "." to avoid conflicting with metadata column names? (Default: FALSE)

Value

A data.frame object. The first column will be named

Sample, and when possible, the rownames will hold the sample name as well. For beta diversity metrics the first two columns will be named Sample1 and Sample2. The remaining columns will be largely dependent on metric, long, and

md. See examples below.

See also

Examples

    library(rbiom)
    
    distill(hmp50, "UniFrac", md=c("==Body Site", "!=Sex")) %>% head()
#>     Sample1 Sample2  Distance     Body Site            Sex
#> 51    HMP01   HMP02 0.1950253 Buccal mucosa Female vs Male
#> 202   HMP02   HMP05 0.2509451 Buccal mucosa Female vs Male
#> 301   HMP01   HMP07 0.1395698 Buccal mucosa Female vs Male
#> 305   HMP05   HMP07 0.1057077 Buccal mucosa Female vs Male
#> 353   HMP03   HMP08 0.3257210        Saliva Female vs Male
#> 354   HMP04   HMP08 0.2328171        Saliva Female vs Male
    
    distill(hmp50, c("Shannon", "OTUs"), md=c("BMI", "Sex")) %>% head()
#>   Sample Depth  Metric Diversity BMI    Sex
#> 1  HMP01  1660 Shannon  1.741153  20 Female
#> 2  HMP02  1371 Shannon  2.587403  23   Male
#> 3  HMP03  1353 Shannon  2.950982  26   Male
#> 4  HMP04  1895 Shannon  3.255785  23   Male
#> 5  HMP05  3939 Shannon  1.462651  24 Female
#> 6  HMP06  4150 Shannon  3.087615  25   Male
    
    distill(hmp50, "Phylum", long=FALSE, md=FALSE)[1:4,1:4]
#>       Actinobacteria Bacteroidetes Cyanobacteria Deinococcus Thermus
#> HMP01             18           276             0                   0
#> HMP02             60           221             0                   0
#> HMP03            126           313             0                   0
#> HMP04            120           218             0                   0
    
    distill(hmp50, "Phylum", long=FALSE, md=c("Age", "Body Site"))[1:4,1:6]
#>       Sample Age     Body Site Actinobacteria Bacteroidetes Cyanobacteria
#> HMP01  HMP01  22 Buccal mucosa             18           276             0
#> HMP02  HMP02  24 Buccal mucosa             60           221             0
#> HMP03  HMP03  28        Saliva            126           313             0
#> HMP04  HMP04  25        Saliva            120           218             0