documentation_default

documentation_default

Arguments

biom

An rbiom-class object, or data coercible with as_rbiom().

mtx

A matrix-like object.

tree

A phylo object representing the phylogenetic relationships of the taxa in biom. Only required when computing UniFrac distances. Default: biom$tree

md

A character vector naming the metadata fields to include in the output data frame, or '.all' to include all metadata fields. Default: '.all'

adiv

Alpha diversity metric(s) to use. Options are: "OTUs", "Shannon", "Chao1", "Simpson", and/or "InvSimpson". Set adiv=".all" to use all metrics. Default: "Shannon"

Multiple values allowed. Non-ambiguous abbreviations are allowed.

bdiv

Beta diversity distance algorithm(s) to use. Options are: "Bray-Curtis", "Manhattan", "Euclidean", "Jaccard", and "UniFrac". For "UniFrac", a phylogenetic tree must be present in biom or explicitly provided via tree=. Default: "Bray-Curtis"

Multiple/abbreviated values allowed.

taxa

Which taxa to display. An integer value will show the top n most abundant taxa. A value 0 <= n < 1 will show any taxa with that mean abundance or greater (e.g. 0.1 implies >= 10%). A character vector of taxa names will show only those named taxa. Default: 6.

ord

Method for reducing dimensionality. Options are:

"UMAP" - Uniform manifold approximation and projection; uwot::umap().

"PCoA" - Principal coordinate analysis; ape::pcoa().

"NMDS" - Nonmetric multidimensional scaling; vegan::metaMDS().

"tSNE" - t-distributed stochastic neighbor embedding; tsne::tsne().

Default: "UMAP"

Multiple values allowed. Non-ambiguous abbreviations are allowed.

weighted

Take relative abundances into account. When weighted=FALSE, only presence/absence is considered. Default: TRUE

Multiple values allowed.

rank

What rank(s) of taxa to display. E.g. "Phylum", "Genus", ".otu", etc. An integer vector can also be given, where 1 is the highest rank, 2 is the second highest, -1 is the lowest rank, -2 is the second lowest, and 0 is the OTU "rank". Run biom$ranks to see all options for a given rbiom object. Default: -1.

lineage

Include all ranks in the name of the taxa. For instance, setting to TRUE will produce Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales. Otherwise the taxa name will simply be Coriobacteriales. You want to set this to TRUE when unc = "asis" and you have taxa names (such as Incertae_Sedis) that map to multiple higher level ranks. Default: FALSE

unc

How to handle unclassified, uncultured, and similarly ambiguous taxa names. Options are:

"singly" - Replaces them with the OTU name.

"grouped" - Replaces them with a higher rank's name.

"drop" - Excludes them from the result.

"asis" - To not check/modify any taxa names.

Default: "singly"

Abbreviations are allowed.

other

Sum all non-itemized taxa into an "Other" taxa. When FALSE, only returns taxa matched by the taxa argument. Specifying TRUE adds "Other" to the returned set. A string can also be given to imply TRUE, but with that value as the name to use instead of "Other". Default: FALSE

sparse

If true, returns a sparse matrix as described by slam::simple_triplet_matrix(), otherwise returns a normal R matrix object. Default: FALSE

p.top

Only display taxa with the most significant differences in abundance. If p.top is >= 1, then the p.top most significant taxa are displayed. If p.top is less than one, all taxa with an adjusted p-value <= p.top are displayed. Recommended to be used in combination with the taxa parameter to set a lower bound on the mean abundance of considered taxa. Default: Inf

y.trans

The transformation to apply to the y-axis. Visualizing differences of both high- and low-abundance taxa is best done with a non-linear axis. Options are:

"sqrt" - square-root transformation

"log1p" - log(y + 1) transformation

NULL - no transformation

These methods allow visualization of both high- and low-abundance taxa simultaneously, without complaint about 'zero' count observations. Default: "sqrt"

flip

Transpose the axes, so that taxa are present as rows instead of columns. Default: FALSE

stripe

Shade every other x position. Default: same as flip

p.label

Minimum adjusted p-value to display on the plot with a bracket.

p.label = 0.05 - Show p-values that are <= 0.05.

p.label = 0 - Don't show any p-values on the plot.

p.label = 1 - Show all p-values on the plot.

If a numeric vector with more than one value is provided, they will be used as breaks for asterisk notation. Default: 0.05

level

The confidence level for calculating a confidence interval. Default: 0.95

caption

Add methodology caption beneath the plot. Default: TRUE

outliers

Show boxplot outliers? TRUE to always show. FALSE to always hide. NULL to only hide them when overlaying a dot or strip chart. Default: NULL

xlab.angle

Angle of the labels at the bottom of the plot. Options are "auto", '0', '30', and '90'. Default: "auto".

k

Number of ordination dimensions to return. Either 2L or 3L. Default: 2L

split.by

Name(s) of metadata columns that the data should be split by prior to any calculations. Default: NULL

dm

A dist-class distance matrix, as returned from bdiv_distmat() or stats::dist(). Required.

groups

A named vector of grouping values. The names should correspond to attr(dm, 'Labels'). Values can be either categorical or numeric. Required.

stat.by

The categorical metadata field defining the statistical groups. Required.

regr

To run a regression analysis, set regr to the numeric metadata field with the "x-axis" values. Leaving regr=NULL will generate boxplot-like statistics; when non-NULL, corrplot-like statistics will be returned. Default: NULL

seed

Random seed for permutations. Default: 0

permutations

Number of random permutations to use. Default: 999

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: "fdr"

depths

Rarefaction depths to show in the plot, or NULL to auto-select. Default: NULL

rline

Where to draw a horizontal line on the plot, intended to show a particular rarefaction depth. Set to TRUE to show an auto-selected rarefaction depth or FALSE to not show a line. Default: NULL

clone

Create a copy of biom before modifying. If FALSE, biom is modified in place as a side-effect. See cloning for use cases. Default: TRUE

labels

Show sample names under each bar. Default: FALSE