Run non-parametric statistics on a data.frame.

Run non-parametric statistics on a data.frame.

Usage

stats_table(
  df,
  stat.by,
  resp = attr(df, "response"),
  regr = NULL,
  test = ifelse(is.null(regr), "means", "trends"),
  model = "lm",
  level = 0.95,
  split.by = NULL,
  p.adj = "fdr"
)

Arguments

df

A data.frame with columns named by stat.by, resp, regr, and split.by. Required.

stat.by

The categorical metadata field defining the statistical groups. Required.

resp

The response (independent) numeric variable, such as taxa abundance or alpha diversity. Default: attr(df, 'response')

regr

To run a regression analysis, set regr to the numeric metadata field with the "x-axis" values. Leaving regr=NULL will generate boxplot-like statistics; when non-NULL, corrplot-like statistics will be returned. Default: NULL

test

The statistic of interest. An overview of options can be found in the "Statistical Tests" section below. One of "predict", "terms", "fit", "means", "trends", "es_means", "es_trends", "pw_means", or "pw_trends". Default: ifelse(is.null(regr), "means", "trends")

model

What type of trend model to fit to the data. Options are: "lm" (linear), "log" (logarithmic), or "gam" (generalized additive). See the "Model Options" section below for additional details. Default: "lm"

level

The confidence level for calculating a confidence interval. Default: 0.95

split.by

Name(s) of metadata columns that the data should be split by prior to any calculations. Default: NULL

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: "fdr".

Value

A tibble data frame with summary statistics.

The R code or generating these statistics is in $code, and the object history is in $history.

Statistical Tests

When regr is NULL:

"means" - Considers all groups at once using the Kruskal-Wallis non-parametric test: stats::kruskal.test().

"pw_means" - Pairwise comparison of stat.by groups using the Wilcox rank sum (aka Mann-Whitney) non-parametric test: stats::wilcox.test().

When regr is not NULL:

"predict" - Augments original data with fitted information. See broom::augment.lm().

"terms" - Summary information about the model's terms, including p-value, r-squared, AIC, BIC, and confidence interval at level. See broom::tidy.lm().

"fit" - Goodness of fit measures, p-values, and more for the overall model. See broom::glance.lm().

"means" - The estimated marginal mean (EMM) for each stat.by group, along with confidence intervals (CI), standard errors (SE), t-ratios, and p-values testing for mean = zero. See emmeans::emmeans() and emmeans::summary.emmGrid().

"trends" - The trendline slope for each stat.by group, along with CI and SE and p-value testing for slope = zero. See emmeans::emtrends() and emmeans::summary.emmGrid().

"pw_means" - Pairwise means. All stat.by groups are compared to each other and the difference in means is estimated along with SE, t-ratios, and p-values testing if the two means are the same. See emmeans::pairs.emmGrid().

"pw_trends" - Pairwise trends. As above, but comparing trendline slopes instead of means.

"es_means" - Effect sizes for pairwise differences of means, including SE and CI. See emmeans::eff_size().

"es_trends" - Effect sizes for pairwise differences of slopes, including SE and CI. See emmeans::eff_size().

Model Options

The predefined options are:

"lm" - Linear model: stats::lm(formula = y ~ x).)

"log" - Logarithmic model: stats::lm(formula = y ~ log(x)).

"gam" - Generalized additive model: mgcv::gam(formula = y ~ s(x, bs = "cs"), method = "REML").

You can alternatively provide a list of length two where the first element is a character vector of length 1 naming a function, and the second element is a list of arguments to pass to that function. One of the function's arguments must be named 'formula'. For example, model = list("stats::lm", list(formula = y ~ x)).

See also

Examples

    library(rbiom)
    
    biom <- sample_rarefy(hmp50)
    
    df <- taxa_table(biom, rank = "Family")
    stats_table(df, stat.by = "Body Site")
#> # Test:     kruskal.test(.abundance ~ `Body Site`).
#> # A tibble: 1 × 5
#>      .n .stat   .df   .p.val   .adj.p
#>   <int> <dbl> <int>    <dbl>    <dbl>
#> 1  3920  148.     4 5.57e-31 5.57e-31
    
    df <- adiv_table(biom)
    stats_table(df, stat.by = "Sex", test = "pw_means")
#> # Test:     wilcox.test(.diversity ~ Sex).
#> # A tibble: 1 × 8
#>   Sex              .n .stat .estimate .lower .upper .p.val .adj.p
#>   <fct>         <dbl> <dbl>     <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1 Female - Male    49   167    -0.781  -1.44 -0.183 0.0149 0.0149
    
    stats_table(df, stat.by = "Sex", regr = "BMI")
#> # Test:     Estimated marginal means of linear trends.
#> # Model:    stats::lm(.diversity ~ BMI * Sex)
#> # A tibble: 2 × 9
#>   Sex    .trend    .se   .df  .lower .upper .t.ratio .p.val .adj.p
#>   <fct>   <dbl>  <dbl> <dbl>   <dbl>  <dbl>    <dbl>  <dbl>  <dbl>
#> 1 Female 0.0653 0.0502    45 -0.0359  0.166    1.30   0.200  0.400
#> 2 Male   0.0546 0.0697    45 -0.0857  0.195    0.784  0.437  0.437