Compute p-values for data vs metadata.

Compute p-values for data vs metadata.

Usage

stats_table(
  biom,
  x,
  y,
  by = NULL,
  adj = "fdr",
  pairwise = FALSE,
  weighted = TRUE,
  digits = 3,
  y.pos = NULL,
  y.pos.facet = "Metric",
  adj.n = NULL
)

Arguments

biom

A BIOM object, as returned from read_biom. Alternatively, a data.frame with column names expected by x, y, and by.

x

The metadata column name of the category to use for comparisons.

y

A character vector with values from one of the following sets:

Alpha Diversity Metrics (one or more)

OTUs, Shannon, Chao1, Simpson, and/or InvSimpson. You may also set y = "Diversity" to get all five metrics.

Beta Diversity Metrics (one only)

manhattan, euclidean, bray-curtis, jaccard, or unifrac. Use in combination with the weighted parameter. Metadata column names can be prefixed with == or != to limit distance calculations to within or between, respectively, those categories. See examples below. Setting y = "Distance" will use unifrac if a phylogenetic tree is present, or bray-curtis otherwise.

Taxa Abundances (one only)

Kingdom, Phylum, Class, Order, Family, Genus, Species, Strain, or OTU. Supported ranks will vary by biom. Run taxa_ranks(biom) to see the available options. Specifying y = "Taxa" will default to the most precise rank possible.

by

Additional metadata columns to group by (e.g. facets in a plot).

adj

Adjustment to use for multiple comparisons. See stats::p.adjust.methods for available options. Default: fdr.

pairwise

If FALSE (the default), one p-value is generated for each by group, or 1 p-value when by = NULL. If TRUE, unique values in x are compared pairwise against each other.

weighted

If TRUE(the default), run beta diversity metrics in weighted mode. Set to FALSE to use unweighted mode.

digits

Round p-values to this many significant figures (default 3). Set to NULL to disable rounding.

y.pos

Add an extra column to the results to indicate the upper most value of each group. Added for ggpubr::geom_bracket(). Options are NULL for no y.pos column (the default), max for the maximum value, box for a box plot's the whisker upper bound, or violin for the highest point of a violin plot.

adj.n

Manually set the number of comparisons to correct for. Only set this if you know what you are doing! Default: NULL

Value

A data.frame with columns .p.val, .adj.p,

adj.sig, as well as columns for tracking x, y, and by categories.

Examples

    library(rbiom)
    
    biom <- rarefy(hmp50)
    biom <- subset(biom, `Body Site` %in% c('Saliva', 'Stool', 'Buccal mucosa'))
#> Error in FUN(X[[i]], ...): object 'Body Site' not found
      
    
    # Overall, do the three body sites different have different Shannon
    #  diversity indices (alpha diversity)? - YES
    stats_table(biom, x = "Body Site", y = "Shannon")
#>    n           Test   .p.val   .adj.p
#> 1 49 Kruskal-Wallis 8.04e-08 8.04e-08
    
    # Which body sites have the most different Shannon diversity
    #  indices? - Buccal mucosa vs. Saliva
    stats_table(biom, x = "Body Site", y = "Shannon", pairwise = TRUE)
#>     n         Test         Group1        Group2   .p.val   .adj.p
#> 1  20 Mann-Whitney Anterior nares        Saliva 1.08e-05 4.33e-05
#> 2  20 Mann-Whitney     Mid vagina        Saliva 1.08e-05 4.33e-05
#> 3  19 Mann-Whitney Anterior nares         Stool 2.17e-05 4.33e-05
#> 4  20 Mann-Whitney  Buccal mucosa        Saliva 2.17e-05 4.33e-05
#> 5  19 Mann-Whitney     Mid vagina         Stool 2.17e-05 4.33e-05
#> 6  19 Mann-Whitney  Buccal mucosa         Stool 1.45e-03 1.88e-03
#> 7  20 Mann-Whitney Anterior nares    Mid vagina 1.50e-03 1.88e-03
#> 8  20 Mann-Whitney  Buccal mucosa    Mid vagina 1.50e-03 1.88e-03
#> 9  19 Mann-Whitney         Saliva         Stool 2.99e-03 3.32e-03
#> 10 20 Mann-Whitney Anterior nares Buccal mucosa 9.71e-01 9.71e-01
    
    # Do males or females have more alpha diversity variation among 
    #  body sites? - Females
    stats_table(biom, x = "Body Site", y = "Shannon", by = "Sex")
#>      Sex  n           Test   .p.val   .adj.p
#> 1 Female 30 Kruskal-Wallis 7.79e-05 0.000156
#> 2   Male 19 Kruskal-Wallis 3.25e-03 0.003250
    
    # Which phylum is most different overall between males and females?
    #  - Saccharibacteria
    stats_table(biom, x = "Sex", y = "Phylum") %>% head()
#>     n         Test .p.val .adj.p
#> 1 637 Mann-Whitney 0.0622 0.0622
    
    # Which phyla is the most differentially abundant between males and 
    #  females on a particular body site? - Tenericutes in Saliva
    stats_table(biom, x = "Sex", y = "Phylum", by = "Body Site") %>% head()
#>        Body Site   n         Test .p.val .adj.p
#> 1         Saliva 130 Mann-Whitney  0.226      1
#> 2 Anterior nares 130 Mann-Whitney  0.530      1
#> 3  Buccal mucosa 130 Mann-Whitney  0.826      1
#> 4          Stool 117 Mann-Whitney  0.912      1
#> 5     Mid vagina 130         <NA>  1.000      1
    
    # Overall, are males and females characterized by their OTU 
    #  abundances? - NO
    stats_table(biom, x = "Sex", y = "unifrac")
#>      n           Test .p.val .adj.p
#> 1 1176 Kruskal-Wallis  0.489  0.489
    
    # What about on a particular body site? - Yes, Buccal mucosa.
    stats_table(biom, x = "==Sex", y = "unifrac", by = "==Body Site")
#> Error in FUN(X[[i]], ...): object '==Body Site' not found
    
    # Recall that for distance metrics such as unifrac, the value of
    #  interest (the distance) is calculated from a pair of samples.
    #  Sometimes both samples will be from males, sometimes both from
    #  females, and sometimes one from each.
    # The '==' prefix on '==Sex' and '==Body Site' indicates that only
    #  pairs from the same sex and same body site should be considered.
    # Using the '!=' prefix does the opposite - considering only sample
    #  pairs from different sexes and body sites.
    
    # Removing the '==' prefix generates the complete set of 
    #  comparisons. Note how the test changes from Mann-Whitney to
    #  Kruskal-Wallis. This is because instead of a two group comparison
    #  ('Male', 'Female') we are now comparing three groups 
    #  ('Male', 'Female', 'Female vs Male').
    stats_table(biom, x = "Sex", y = "unifrac", by = "Body Site")
#>                          Body Site   n           Test  .p.val .adj.p
#> 1      Buccal mucosa vs Mid vagina 100   Mann-Whitney 0.00228 0.0239
#> 2                    Buccal mucosa  45 Kruskal-Wallis 0.00319 0.0239
#> 3                  Saliva vs Stool  90 Kruskal-Wallis 0.15100 0.6110
#> 4          Buccal mucosa vs Saliva 100 Kruskal-Wallis 0.16300 0.6110
#> 5              Mid vagina vs Stool  90   Mann-Whitney 0.45700 1.0000
#> 6             Mid vagina vs Saliva 100   Mann-Whitney 0.55600 1.0000
#> 7                   Anterior nares  45 Kruskal-Wallis 0.64400 1.0000
#> 8           Buccal mucosa vs Stool  90 Kruskal-Wallis 0.79700 1.0000
#> 9                           Saliva  45 Kruskal-Wallis 0.80700 1.0000
#> 10         Anterior nares vs Stool  90 Kruskal-Wallis 0.94300 1.0000
#> 11                           Stool  36 Kruskal-Wallis 0.95800 1.0000
#> 12        Anterior nares vs Saliva 100 Kruskal-Wallis 0.98200 1.0000
#> 13 Anterior nares vs Buccal mucosa 100 Kruskal-Wallis 0.99000 1.0000
#> 14    Anterior nares vs Mid vagina 100   Mann-Whitney 0.99700 1.0000
#> 15                      Mid vagina  45           <NA> 1.00000 1.0000