Subset an rbiom object by sample names or metadata.

Subset an rbiom object by sample names or metadata.

Usage

# S3 method for class 'rbiom'
subset(x, subset, clone = TRUE, ...)

# S3 method for class 'rbiom'
biom[i]

# S3 method for class 'rbiom'
na.omit(object, fields = ".all", clone = TRUE, ...)

Arguments

x

An rbiom object, such as from as_rbiom().

subset

Logical expression for rows to keep. See base::subset().

clone

Create a copy of biom before modifying. If FALSE, biom is modified in place as a side-effect. See speed ups for use cases. Default: TRUE

...

Not used.

biom

An rbiom object, such as from as_rbiom().

i

The sample names to keep. Or a logical/integer vector indicating which sample names from biom$samples to keep.

object

An rbiom object, such as from as_rbiom().

fields

Which metadata field(s) to check for NAs, or ".all" to check all metadata fields.

Value

An rbiom object.

See also

Examples

    library(rbiom) 
    
    # Subset to specific samples
    biom <- hmp50[c('HMP20', 'HMP42', 'HMP12')]
    biom$metadata
#> # A tibble: 3 × 5
#>   .sample   Age   BMI `Body Site` Sex   
#>   <chr>   <dbl> <dbl> <fct>       <fct> 
#> 1 HMP20      27    22 Stool       Female
#> 2 HMP42      34    19 Mid vagina  Female
#> 3 HMP12      35    26 Stool       Male  
    
    # Subset according to metadata (using base::subset)
    biom <- subset(hmp50, `Body Site` %in% c('Saliva') & Age < 25)
    biom$metadata
#> # A tibble: 3 × 5
#>   .sample   Age   BMI `Body Site` Sex   
#>   <chr>   <dbl> <dbl> <fct>       <fct> 
#> 1 HMP27      24    30 Saliva      Female
#> 2 HMP28      23    19 Saliva      Female
#> 3 HMP30      24    21 Saliva      Female
    
    # Remove samples with NA metadata values
    biom <- mutate(hmp50, BS2 = na_if(`Body Site`, 'Saliva'))
#> Error in dplyr::mutate(.data = biom$metadata, ...): ℹ In argument: `BS2 = na_if(`Body Site`, "Saliva")`.
#> Caused by error in `na_if()`:
#> ! could not find function "na_if"
    biom$metadata
#> # A tibble: 3 × 5
#>   .sample   Age   BMI `Body Site` Sex   
#>   <chr>   <dbl> <dbl> <fct>       <fct> 
#> 1 HMP27      24    30 Saliva      Female
#> 2 HMP28      23    19 Saliva      Female
#> 3 HMP30      24    21 Saliva      Female
    biom <- na.omit(biom)
    biom$metadata
#> # A tibble: 3 × 5
#>   .sample   Age   BMI `Body Site` Sex   
#>   <chr>   <dbl> <dbl> <fct>       <fct> 
#> 1 HMP27      24    30 Saliva      Female
#> 2 HMP28      23    19 Saliva      Female
#> 3 HMP30      24    21 Saliva      Female