Subset samples using the BIOM's metadata or taxonomy.

Subset samples using the BIOM's metadata or taxonomy.

# S3 method for BIOM
subset(
  x,
  expr,
  env = parent.frame(),
  drop.na = TRUE,
  refactor = TRUE,
  fast = FALSE
)

Arguments

x

A BIOM object, as returned from read_biom.

expr

Logical expression to run on the metadata or taxonomy (not both) to identify samples or taxa to retain.

env

The environment to search for variables used in expr. (Default: parent.frame())

drop.na

When expr is e.g. Age > 30, should !is.na(Age) be automatically applied too?

refactor

When expr is e.g. `Body Site` %in% c("Stool", "Saliva"), should `Body Site` be redefined as factor(`Body Site`, levels=c('Stool', 'Saliva'))? Applies to non-numeric metadata only.

fast

Should subsetting the phylogenetic tree and sequences be skipped? These slow steps are often not necessary. (Default: FALSE)

Value

A BIOM object.

Taxonomic abundance filtering

For taxonomic subsetting, several functions are added or overridden to behave as expected within the subsetting expression. They are:

mean(), median(), min(), max(), n(), count(), percent(), rank(), and apply().

Therefore you can write subset(hmp50, mean(Genus) >= 0.1) and the returned BIOM object will contain only the genera that average at least 10% relative abundance across all the samples.

If you want only orders that are present in three or more samples, you can do: subset(hmp50, count(Order) >= 3). To require presence in 25% of samples, you'd use: subset(hmp50, percent(Order) >= 0.25).

Both count() and percent() have default arguments of gt=0, le=1, ge=NULL, lt=NULL, which can be overridden to find, e.g., which genera comprise at least 2% of the community in 10% or more of the samples: subset(hmp50, percent(Genus, ge=0.02) >= 0.10).

gt = greater than, ge = greater than or equal to. lt/le similarly with 'less than'.

To keep only the top 5 most abundant genera (based on mean), run: subset(hmp50, rank(Genus) <= 5).

apply() allows you to run any function on a per-taxon basis. For example, to filter genera by the root mean square of their relative abundances: rms <- function(x) sqrt(mean(x^2)); subset(hmp50, apply(Genus, rms) >= 0.1).

If you prefer to work on the raw values (e.g. read counts) instead of relative abundances, set apply(..., raw = TRUE). For instance: subset(hmp50, apply(Genus, mean, raw=TRUE) >= 100).

See also

Examples

  if (FALSE) {
    library(rbiom)
    
    subset(hmp50, `Body Site` %in% c("Saliva", "Stool"))
    subset(hmp50, Age < 25 & BMI > 22)
    subset(hmp50, Phylum %in% c("Firmicutes", "Actinobacteria"))
    subset(hmp50, mean(Genus) > 0.1)
    subset(hmp50, rank(Genus) <= 5)
    subset(hmp50, a == b, list(a = as.name("Body Site"), b ="Saliva"))
 }