Test taxa abundances for significant differences.

Test taxa abundances for significant differences.

Usage

taxa_stats(
  biom,
  stat.by,
  rank = -1,
  taxa = 6,
  split.by = NULL,
  lineage = FALSE,
  unc = "singly",
  other = FALSE,
  trans = "none",
  test = "means",
  p.adj = "fdr"
)

Arguments

biom

An rbiom-class object, or data coercible with as_rbiom().

stat.by

The categorical metadata field defining the statistical groups. Required.

rank

What rank(s) of taxa to display. E.g. "Phylum", "Genus", ".otu", etc. An integer vector can also be given, where 1 is the highest rank, 2 is the second highest, -1 is the lowest rank, -2 is the second lowest, and 0 is the OTU "rank". Run biom$ranks to see all options for a given rbiom object. Default: -1.

taxa

Which taxa to display. An integer value will show the top n most abundant taxa. A value 0 <= n < 1 will show any taxa with that mean abundance or greater (e.g. 0.1 implies >= 10%). A character vector of taxa names will show only those named taxa. Default: 6.

split.by

Name(s) of metadata columns that the data should be split by prior to any calculations. Default: NULL

lineage

Include all ranks in the name of the taxa. For instance, setting to TRUE will produce Bacteria; Actinobacteria; Coriobacteriia; Coriobacteriales. Otherwise the taxa name will simply be Coriobacteriales. You want to set this to TRUE when unc = "asis" and you have taxa names (such as Incertae_Sedis) that map to multiple higher level ranks. Default: FALSE

unc

How to handle unclassified, uncultured, and similarly ambiguous taxa names. Options are:

"singly" - Replaces them with the OTU name.

"grouped" - Replaces them with a higher rank's name.

"drop" - Excludes them from the result.

"asis" - To not check/modify any taxa names.

Default: "singly"

Abbreviations are allowed.

other

Sum all non-itemized taxa into an "Other" taxa. When FALSE, only returns taxa matched by the taxa argument. Specifying TRUE adds "Other" to the returned set. A string can also be given to imply TRUE, but with that value as the name to use instead of "Other". Default: FALSE

trans

Transformation to apply. Options are: c("none", "rank", "log", "log1p", "sqrt"). "rank" is useful for correcting for non-normally distributions before applying regression statistics. Default: "none"

p.adj

Method to use for multiple comparisons adjustment of p-values. Run p.adjust.methods for a list of available options. Default: "fdr"

Value

A tibble data frame with summary statistics.

The R code or generating these statistics is in $code.

See also

Examples

    library(rbiom)
    
    biom <- rarefy(hmp50)
    
    taxa_stats(biom, stat.by = "Body Site", rank = "Family")
#> # Test:     kruskal.test(.Family ~ `Body Site`).
#> # A tibble: 6 × 6
#>   .taxa                 .n .stat   .df       .p.val       .adj.p
#>   <fct>              <int> <dbl> <int>        <dbl>        <dbl>
#> 1 Lactobacillaceae      49  37.4     4 0.000000148  0.000000296 
#> 2 Streptococcaceae      49  42.8     4 0.0000000114 0.0000000685
#> 3 Bacteroidaceae        49  26.9     4 0.0000204    0.0000204   
#> 4 Corynebacteriaceae    49  29.6     4 0.00000584   0.00000701  
#> 5 Pasteurellaceae       49  40.3     4 0.0000000372 0.000000112 
#> 6 Veillonellaceae       49  32.0     4 0.00000191   0.00000286