Write data and summary information to a Microsoft Excel-compatible workbook.

Write data and summary information to a Microsoft Excel-compatible workbook.

Usage

write_xlsx(biom, outfile, depth = NULL, seed = 0, unc = "asis")

Arguments

biom

The BIOM object to save to the file.

outfile

Path to the output xlsx file.

depth

Depth to rarefy to. See rarefy function for details. depth = NULL auto-selects a rarefaction level. depth = 0 disables rarefaction. Only use depth with BIOM files of type 'OTU table' and integer count values.

seed

Random seed to use in rarefying. See rarefy function for details.

unc

How to handle unclassified, uncultured, and similarly ambiguous taxa names. Default: "asis"

"asis" - Don't check/modify any taxa names.

"singly" - Replace with "Unc. <OTU ID>".

"grouped" - Replace with "Unc. <Higher Rank>".

"drop" - Don't include in the returned matrix.

Value

On success, returns NULL invisibly.

Note

Any data frame, matrix, or dist attributes on biom will be included as separate worksheets. An attribute named 'Reads Per Step' is treated specially and merged with the usual 'Reads Per Sample' tab.

Examples

    library(rbiom)
    
    biom <- select(hmp50, 1:10) %>% rarefy()
    
    attr(biom, "Weighted UniFrac")   <- unifrac(biom)
    attr(biom, "Unweighted Jaccard") <- bdiv_distmat(biom, 'jaccard', weighted=FALSE)
    
    outfile <- tempfile(fileext = ".xlsx")
    write_xlsx(biom, outfile)